### 2020-02-08 #2086 @whedon
[PRE REVIEW]: fastSF: A parallel code for computing the structure functions of turbulence
**Submitting author:** @ShubhadeepSadhukhan1993 (<a href="http://orcid.org/0000-0002-7278-5041">Shubhadeep Sadhukhan</a>)
**Repository:** <a href="https://github.com/ShubhadeepSadhukhan1993/fastSF" target ="_blank">https://github.com/ShubhadeepSadhukhan1993/fastSF</a>
**Version:** v1.0.0
