### 2020-02-09 #2086 @jedbrown
Thanks for your submission. In reading the paper and checking the code, I'm concerned that the implementation strategy here is highly suboptimal, and should really be improved if it is to be widely used. At its core, the straightforward use of six nested loops to compute differences between values at every pair of points in a 3D grid is similar to that of using naive loops for matrix-matrix product in terms of having very poor cache reuse and vectorization. My calculations suggest your test on Shaheen II ran at around 5% of peak, but this operation should be floating point limited, and a high fraction of peak should be attainable with tiling and better vectorization (you're asking the compiler to do quite a lot with how the code is written at present). I also think it would be highly preferable if your parallel implementation did not perform independent IO and full storage of global data on each node. I understand this may not be a bottleneck now because everything else is so slow. I'm also looking for a value-add because a user writing the naive nested loops and handful of lines inside is likely easier than conforming to your IO standard and using your software. I would also expect there are fast approximate methods that run much faster -- do you know if such methods exist in the literature?

Also, please note that your software has a hard dependency on H5SI, which I can't find anywhere except in the tarball of unknown provenance on the non-https website. That tarball contains no license file, but has GPL comments in source files. GPL is not compatible with the BSD license in your project. We cannot evaluate your project further with this license incompatibility.
