### 2020-02-11 #2086 @danielskatz
👋 @openjournals/dev - something happened in the initial comment in this issue where the author's github handle (I think ShubhadeepSadhukhan1993) didn't get added correctly. I could just edit the comment but I would prefer you to check the database to make sure this is correct internally.
