### 2020-02-12 #2086 @shashwat-bhattacharya
@jedbrown We thank you for your comments and feedback. We are modifying the code as per your recommendations. Further, the authors of H5SI are in the process of implementing BSD license for H5SI software. We will get back to you once these modifications are done.
