### 2020-02-16 9342761
Implemented vectorization with three loops instead of 6; removed OpenMP
