+++ src/Kolmogorov41.cc
@@ -1,10 +1,10 @@
 /********************************************************************************************************************************************
  * Kolmogorov41
- * 
+ *
  * Copyright (C) 2020, Mahendra K. Verma
  *
  * All rights reserved.
- * 
+ *
  * Redistribution and use in source and binary forms, with or without
  * modification, are permitted provided that the following conditions are met:
  *     1. Redistributions of source code must retain the above copyright
@@ -15,7 +15,7 @@
  *     3. Neither the name of the copyright holder nor the
  *        names of its contributors may be used to endorse or promote products
  *        derived from this software without specific prior written permission.
- * 
+ *
  * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS" AND
  * ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED
  * WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE
@@ -33,9 +33,9 @@
 /*! \file Kolmogorov41.cc
  *
  *  \brief Code to compute structure functions using velocity or scalar field data.
- * 
- *  \author Shubhadeep Sadhukhan, Shashwat Bhattacharya, Mahendra K. Verma
- *  \date Feb 2020
+ *
+ *  \author Shubhadeep Sadhukhan, Shashwat Bhattacharya
+ *  \date Jan 2020
  *  \copyright New BSD License
  *
  ********************************************************************************************************************************************
@@ -68,74 +68,82 @@ void VECTOR_TEST_CASE_2D();
 void SCALAR_TEST_CASE_2D();
 void SCALAR_TEST_CASE_3D();
 void compute_time_elapsed(timeval, timeval, double&);
-double powInt(double, int);
+//double powInt(double, int);
 void read_3D(Array<double,3>, string, string);
-void SFunc2D(Array<double,2>, Array<double,2>, Array<double,2>&, Array<double,2>&, Array<double,2>&);
-void SFunc2D(Array<double,2>, Array<double,2>, Array<double,2>&, Array<double,2>&, Array<double,2>&, Array<double,3>&, Array<double,3>&, Array<double,3>&);
-void SFunc_long_2D(Array<double,2>, Array<double,2>, Array<double,2>&, Array<double,2>&);
-void SFunc_long_2D(Array<double,2>, Array<double,2>, Array<double,2>&, Array<double,2>&, Array<double,3>&, Array<double,3>&);
-void SFunc3D(Array<double,3>, Array<double,3>, Array<double,3>, Array<double,2>&, Array<double,2>&, Array<double,2>&);
-void SFunc3D(Array<double,3>, Array<double,3>, Array<double,3>, Array<double,2>&, Array<double,2>&, Array<double,2>&, Array<double,4>&, Array<double,4>&, Array<double,4>&);
-void SFunc_long_3D(Array<double,3>, Array<double,3>, Array<double,3>, Array<double,2>&, Array<double,2>&);
-void SFunc_long_3D(Array<double,3>, Array<double,3>, Array<double,3>, Array<double,2>&, Array<double,2>&, Array<double,4>&, Array<double,4>&);
+
+void SFunc2D(Array<double,2>, Array<double,2>, Array<double,2>&, Array<double,2>&);
+void SFunc2D(Array<double,2>, Array<double,2>, Array<double,2>&, Array<double,2>&, Array<double,2>&, Array<double,3>&, Array<double,3>&);
+
[…]
