+++ in/para.yaml
@@ -1,29 +1,28 @@
 #PARAMETERS FOR COMPUTING THE STRUCTURE FUNCTIONS"
 
 program:
-    #Please enter "true" for computing structure functions as function of lx, ly (for 3D), and lz along with l. If "false is selected, only structure functions as function of l
+    #Please select "true" for computing structure functions as function of lx, ly (for 3D), and lz along with l. If "false is selected, only structure functions as function of l
     #will be computed.
-    grid_switch: false
+    grid_switch: true
 
-    #Please enter "true" for computing scalar structure function, "false" for computing velocity structure function:
+    #Please select "true" for computing scalar structure function, "false" for computing velocity structure function:
     scalar_switch: false
   
-    #Please enter "true" for 2D operations, "false" for 3D operations:
+    #Please select "true" for 2D operations, "false" for 3D operations:
     2D_switch : false
 
-    #Please enter "true" for computing only the longitudinal structure functions, "false" for computing both the transverse and longitudinal structure functions:
-    Only_longitudinal: true
+    #Please select "true" for computing only the longitudinal structure functions, "false" for computing both the transverse and longitudinal structure functions:
+    Only_longitudinal: false
 
-    #Please enter the number of OpenMP processors
-    Number_of_OpenMP_processors: 16
+    Number_of_OpenMP_processors: 1
 
 #Please specify the number of grid points. 
 #Note: Nx - number of points in the x direction, Ny - number of points in the y-direction, Nz - number of points in the z direction.
 #For 2D, provide Nx and Nz.
 grid :
-    Nx : 32
-    Ny : 32
-    Nz : 32
+    Nx : 64
+    Ny : 64
+    Nz : 64
 
         
 #Please specify the domain dimensions. 
@@ -37,8 +36,8 @@ domain_dimension :
 
 #Please provide the starting order (q1) and the ending order (q2)
 structure_function :
-    q1 : 3
-    q2 : 3
+    q1 : 1
+    q2 : 10
 
 #Please enter "true" only if you want to run a test case. WARNING: For test cases, the input fields will be generated by the code. The code will ignore
 # the hdf5 files in the "in" folder. Further,the grid_switch will be automatically set to "true". Thus, the entries against "grid_switch" and "field_procedure" 
