+++ in/para.yaml
@@ -20,9 +20,9 @@ program:
 #Note: Nx - number of points in the x direction, Ny - number of points in the y-direction, Nz - number of points in the z direction.
 #For 2D, provide Nx and Nz.
 grid :
-    Nx : 64
-    Ny : 64
-    Nz : 64
+    Nx : 32
+    Ny : 32
+    Nz : 32
 
         
 #Please specify the domain dimensions. 
@@ -36,8 +36,8 @@ domain_dimension :
 
 #Please provide the starting order (q1) and the ending order (q2)
 structure_function :
-    q1 : 1
-    q2 : 10
+    q1 : 2
+    q2 : 3
 
 #Please enter "true" only if you want to run a test case. WARNING: For test cases, the input fields will be generated by the code. The code will ignore
 # the hdf5 files in the "in" folder. Further,the grid_switch will be automatically set to "true". Thus, the entries against "grid_switch" and "field_procedure" 
