### 2020-02-16 420249e
Loops are upto Nx/2, Ny/2, and Nz/2 instead of Nx, Ny, and Nz
