+++ src/Kolmogorov41.cc
@@ -57,7 +57,7 @@ using namespace std;
 using namespace blitz;
 
 //Function declarations
-void Read_para(); //Input from the user
+void Read_para(); 
 void write_2D(Array<double,2>, string);
 void write_3D(Array<double,3>, string, int);
 void write_4D(Array<double,4>, string, int);
@@ -68,19 +68,19 @@ void VECTOR_TEST_CASE_2D();
 void SCALAR_TEST_CASE_2D();
 void SCALAR_TEST_CASE_3D();
 void compute_time_elapsed(timeval, timeval, double&);
-//double powInt(double, int);
+
 void read_3D(Array<double,3>, string, string);
 
-void SFunc2D(Array<double,2>, Array<double,2>, Array<double,2>&, Array<double,2>&);
+
 void SFunc2D(Array<double,2>, Array<double,2>, Array<double,2>&, Array<double,2>&, Array<double,2>&, Array<double,3>&, Array<double,3>&);
 
-void SFunc_long_2D(Array<double,2>, Array<double,2>, Array<double,2>&);
+
 void SFunc_long_2D(Array<double,2>, Array<double,2>, Array<double,2>&, Array<double,2>&, Array<double,3>&);
 
-void SFunc3D(Array<double,3>, Array<double,3>, Array<double,3>, Array<double,2>&, Array<double,2>&);
+
 void SFunc3D(Array<double,3>, Array<double,3>, Array<double,3>, Array<double,2>&, Array<double,2>&, Array<double,2>&, Array<double,4>&, Array<double,4>&);
 
-void SFunc_long_3D(Array<double,3>, Array<double,3>, Array<double,3>, Array<double,2>&);
+
 void SFunc_long_3D(Array<double,3>, Array<double,3>, Array<double,3>, Array<double,2>&, Array<double,2>&, Array<double,4>&);
 
 void Read_Init(Array<double,2>&, Array<double,2>&);
@@ -90,10 +90,10 @@ void Read_Init(Array<double,3>&);
 
 double powInt(double, int);
 
-void SF_scalar_3D(Array<double,3>, Array<double,2>&);
+
 void SF_scalar_3D(Array<double,3>, Array<double,2>&, Array<double,2>&, Array<double,4>&);
 
-void SF_scalar_2D(Array<double,2>, Array<double,2>&);
+
 void SF_scalar_2D(Array<double,2>, Array<double,2>&, Array<double,2>&, Array<double,3>&);
 void compute_index_list(Array<int, 2>&);
 
@@ -242,17 +242,6 @@ bool two_dimension_switch;
  */
 bool scalar_switch;
 
-/**
- ********************************************************************************************************************************************
- * \brief    This variable decides whether the structure functions as function of \f$ (l_x, l_y, l_z) \f$, or of \f$ (l_x, l_z) \f$ in case of
- *           2D, needs to be calculated in addition to the structure functions as function of \f$ l \f$.
- *
- * If the value is TRUE, then the structure functions as function of \f$ (l_x, l_y, l_z) \f$ or \f$  (l_x, l_z) \f$ will be computed in addtion to the structure
- * functions as function of \f$ l \f$.
- ********************************************************************************************************************************************
- */
-bool grid_switch;
-
 
 /**
  ********************************************************************************************************************************************
@@ -284,12 +273,6 @@ int Ny;
  */
[…]
