+++ in/para.yaml
@@ -1,10 +1,6 @@
 #PARAMETERS FOR COMPUTING THE STRUCTURE FUNCTIONS"
 
 program:
-    #Please select "true" for computing structure functions as function of lx, ly (for 3D), and lz along with l. If "false is selected, only structure functions as function of l
-    #will be computed.
-    grid_switch: true
-
     #Please select "true" for computing scalar structure function, "false" for computing velocity structure function:
     scalar_switch: false
   
@@ -13,8 +9,6 @@ program:
 
     #Please select "true" for computing only the longitudinal structure functions, "false" for computing both the transverse and longitudinal structure functions:
     Only_longitudinal: false
-
-    Number_of_OpenMP_processors: 1
 
 #Please specify the number of grid points. 
 #Note: Nx - number of points in the x direction, Ny - number of points in the y-direction, Nz - number of points in the z direction.
@@ -36,8 +30,8 @@ domain_dimension :
 
 #Please provide the starting order (q1) and the ending order (q2)
 structure_function :
-    q1 : 2
-    q2 : 3
+    q1 : 1
+    q2 : 10
 
 #Please enter "true" only if you want to run a test case. WARNING: For test cases, the input fields will be generated by the code. The code will ignore
 # the hdf5 files in the "in" folder. Further,the grid_switch will be automatically set to "true". Thus, the entries against "grid_switch" and "field_procedure" 
