+++ src/Kolmogorov41.cc
@@ -371,7 +371,13 @@ double Lz;
  */
 int P;
 
-
+/**
+ ********************************************************************************************************************************************
+ * \brief   This variable stores the number of the MPI process along x axis.
+ *
+ ********************************************************************************************************************************************
+ */
+int px;
 //$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$//
 
 
@@ -391,6 +397,7 @@ int main(int argc, char *argv[]) {
     MPI_Init(NULL, NULL);
     MPI_Comm_rank(MPI_COMM_WORLD, &rank_mpi);
     MPI_Comm_size(MPI_COMM_WORLD, &P);
+
     //Initiallizing h5si
     h5::init();
     timeval start_pt, end_pt, start_t, end_t;
@@ -398,12 +405,14 @@ int main(int argc, char *argv[]) {
     double elapsedt=0.0;
     double elapsepdt=0.0;
     Read_para();
+
+    /*
     if (Nx/2%P!=0){
         if (rank_mpi==0){
             cout<<"proccesor number should be less or equal to Nx/4 and some power of 2\n Exiting the code\n";
             exit(1);
         }
-    }
+    }*/
     //Resizing the input fields
 
     if(two_dimension_switch){
@@ -658,6 +667,58 @@ int main(int argc, char *argv[]) {
 
 
 /**
+*************
+*************
+*/
+void get_rank(int rank, int py, int& rankx, int& ranky){
+    ranky=rank%py;
+    rankx=(rank-ranky)/py;
+}
+void compute_index_list(Array<int,1>& index_list, int Nx, int px, int rank){
+    int list_size=Nx/px;
+    index_list.resize(list_size);
+    for (int i=0; i<list_size; i+=2){
+        index_list(i)=rank+i*px;
+        if (px!=Nx) {
+          index_list(i+1)=Nx-1-index_list(i);
+        }
+    }
+}
+
+
+void compute_load(Array<int,3> index_list, int rank_id){
+    Array<int,1> load(index_list(Range::all(),0,0).size());
+    load=(Nx-index_list(Range::all(),0,rank_id))*(Ny-index_list(Range::all(),1,rank_id));
+    cout<< sum(load)<<endl;
+}
+
+void compute_index_list(Array<int,3>& index_list, int Nx, int Ny){
+    int list_size=(Nx*Ny)/(4*P);
+    index_list.resize(list_size,2,P);
+
+    int py=(P/px);
+    Array<int,1> x;
+    Array<int,1> y;
+    
+    int rankx,ranky;
+    int nx=Nx/(2*px),ny=Ny/(2*py);
+
+    
+    for (int rank_id=0; rank_id<P; rank_id++){
+        get_rank(rank_id, py, rankx, ranky);
+        compute_index_list(x, Nx/2, px, rankx);
+        compute_index_list(y, Ny/2, py, ranky);
+        for (int i=0; i<nx; i++){
+            index_list(Range(ny*i,(i+1)*ny-1),0,rank_id)=x(i);
[…]
