+++ in/para.yaml
@@ -10,7 +10,7 @@ program:
     #Please select "true" for computing only the longitudinal structure functions, "false" for computing both the transverse and longitudinal structure functions:
     Only_longitudinal: false
 
-number_of_proc_along_x: 4 
+number_of_proc_along_x: 1 
 
 #Please specify the number of grid points. 
 #Note: Nx - number of points in the x direction, Ny - number of points in the y-direction, Nz - number of points in the z direction.
@@ -32,8 +32,8 @@ domain_dimension :
 
 #Please provide the starting order (q1) and the ending order (q2)
 structure_function :
-    q1 : 1
-    q2 : 10
+    q1 : 3
+    q2 : 3
 
 #Please enter "true" only if you want to run a test case. WARNING: For test cases, the input fields will be generated by the code. The code will ignore
 # the hdf5 files in the "in" folder. Further,the grid_switch will be automatically set to "true". Thus, the entries against "grid_switch" and "field_procedure" 
