### 2020-02-17 acc9607
Removed the nodal arrays; Abstraction in the main function; SF(l) removed
