+++ src/Kolmogorov41.cc
@@ -72,16 +72,16 @@ void compute_time_elapsed(timeval, timeval, double&);
 void read_3D(Array<double,3>, string, string);
 
 
-void SFunc2D(Array<double,2>, Array<double,2>, Array<double,2>&, Array<double,2>&, Array<double,2>&, Array<double,3>&, Array<double,3>&);
+void SFunc2D(Array<double,2>, Array<double,2>);
 
 
-void SFunc_long_2D(Array<double,2>, Array<double,2>, Array<double,2>&, Array<double,2>&, Array<double,3>&);
+void SFunc_long_2D(Array<double,2>, Array<double,2>);
 
 
-void SFunc3D(Array<double,3>, Array<double,3>, Array<double,3>, Array<double,2>&, Array<double,2>&, Array<double,2>&, Array<double,4>&, Array<double,4>&);
+void SFunc3D(Array<double,3>, Array<double,3>, Array<double,3>);
 
 
-void SFunc_long_3D(Array<double,3>, Array<double,3>, Array<double,3>, Array<double,2>&, Array<double,2>&, Array<double,4>&);
+void SFunc_long_3D(Array<double,3>, Array<double,3>, Array<double,3>);
 
 void Read_Init(Array<double,2>&, Array<double,2>&);
 void Read_Init(Array<double,3>&, Array<double,3>&, Array<double,3>&);
@@ -91,11 +91,18 @@ void Read_Init(Array<double,3>&);
 double powInt(double, int);
 
 
-void SF_scalar_3D(Array<double,3>, Array<double,2>&, Array<double,2>&, Array<double,4>&);
+void SF_scalar_3D(Array<double,3>);
 
 
-void SF_scalar_2D(Array<double,2>, Array<double,2>&, Array<double,2>&, Array<double,3>&);
+void SF_scalar_2D(Array<double,2>);
 void compute_index_list(Array<int, 2>&);
+
+void Read_fields();
+void resize_SFs();
+void calc_SFs();
+void write_SFs();
+void test_cases();
+
 
 
 /**
@@ -150,37 +157,6 @@ Array<double,2> V3_2D;
 
 /**
  ********************************************************************************************************************************************
- * \brief   2D array storing the computed longitudinal structure functions or scalar structure functions and their corresponding orders.
- *
- *          The dimensions of this array is \f$N_l \times p\f$, where \f$N_l\f$ is the number of points along the diagonal of the computation
- *          domain, and \f$p \f$ is the number of orders of the structure functions to be computed. Thus, each column of the array represents
- *          a particular order of the structure functions.
- ********************************************************************************************************************************************
- */
-Array<double,2> SF;
-
-/**
- ********************************************************************************************************************************************
- * \brief   2D array storing the computed transverse structure functions and their corresponding orders.
-
- *          The dimensions of this array is \f$N_l \times p\f$, where \f$N_l\f$ is the number of points along the diagonal of the computation
- *          domain, and \f$p \f$ is the number of orders of the structure functions to be computed. Thus, each column of the array represents
- *          a particular order of the structure functions.
- ********************************************************************************************************************************************
- */
-Array<double,2> SF_perp;
-
-
-/**
- ********************************************************************************************************************************************
- * \brief   2D array for storing the values to divide SF and SF_perp for averaging. 
- *
[…]
