+++ in/para.yaml
@@ -32,8 +32,8 @@ domain_dimension :
 
 #Please provide the starting order (q1) and the ending order (q2)
 structure_function :
-    q1 : 3
-    q2 : 3
+    q1 : 1
+    q2 : 10
 
 #Please enter "true" only if you want to run a test case. WARNING: For test cases, the input fields will be generated by the code. The code will ignore
 # the hdf5 files in the "in" folder. Further,the grid_switch will be automatically set to "true". Thus, the entries against "grid_switch" and "field_procedure" 
