### 2020-02-17 362940d
Removed Nr and its calculations as it is no longer required
