+++ paper.md
@@ -25,7 +25,7 @@ affiliations:
  - name: Department of Mechanical Engineering, Indian Institute of Technology Kanpur 208016, India
    index: 2
 
-date: 06 February 2020
+date: 18 February 2020
 
 bibliography: paper.bib
 
@@ -33,9 +33,9 @@ bibliography: paper.bib
 
 # Summary
 
-Turbulence is a complex phenomenon in fluid dynamics involving nonlinear interactions between multiple scales. Structure function is a popular diagnostics tool to study the statistical properites of turbulent flows [@Kolmogorov:Dissipation; @Kolmogorov:Structure; @Frisch:book]. Some of the earlier works comprising of such analysis are those of @Gotoh:PF2002, @Kaneda:PF2003, and @Ishihara:ARFM2009 for three-dimensional homogeneous isotropic turbulence; @Yeung:PF2005 and @Ray:NJP2008 for passive scalar turbulence; @Biferale:NJP2004 for two-dimensional turbulence; and @Kunnen:PRE2008 and @Kaczorowski:JFM2013 for turbulent thermal convection. Structure functions are two-point statistical quantities; thus, an accurate computation of these quantities requires averaging over many points. However, incorporation of a large number of points makes the computations very expensive and challenging. Therefore, we require an efficient parallel code for accurate computation of structure functions. In this paper, we describe the design and validation of the results of ``Kolmogorov41``, a hybrid parallel code to compute velocity and scalar structure functions. 
+Turbulence is a complex phenomenon in fluid dynamics involving nonlinear interactions between multiple scales. Structure function is a popular diagnostics tool to study the statistical properites of turbulent flows [@Kolmogorov:Dissipation; @Kolmogorov:Structure; @Frisch:book]. Some of the earlier works comprising of such analysis are those of @Gotoh:PF2002, @Kaneda:PF2003, and @Ishihara:ARFM2009 for three-dimensional homogeneous isotropic turbulence; @Yeung:PF2005 and @Ray:NJP2008 for passive scalar turbulence; @Biferale:NJP2004 for two-dimensional turbulence; and @Kunnen:PRE2008, @Kaczorowski:JFM2013, and @Bhattacharya:PF2019 for turbulent thermal convection. Structure functions are two-point statistical quantities; thus, an accurate computation of these quantities requires averaging over many points. However, incorporation of a large number of points makes the computations very expensive and challenging. Therefore, we require an efficient parallel code for accurate computation of structure functions. In this paper, we describe the design and validation of the results of ``fastSF``, a hybrid parallel code to compute velocity and scalar structure functions. 
 
- ``Kolmogorov41``, written in C++, employs a combination of distributed (MPI) and shared (OpenMP) memory parallelization [@Pacheco:book:PP] to compute the structure functions for a given velocity or scalar field. In this code, the entire data is shared in all the MPI processes, thus eliminating the need of communication between the processors. Thus, we save considerable time that would otherwise be spent on communication. The user has a choice on the type (scalar or vector) and the dimensions of the fields to be read by the code, and the range of the orders of the structure functions to be computed. The code writes the computed structure functions to hdf5 files that can be further processed by the user. We remark that our code has been used by @Bhattacharya:PF2019 for analysing the longitudinal velocity structure functions of turbulent thermal convection.
[…]
