### 2020-02-18 0ea7ab7
Update paper.md
Updated "Design of the Code"
