+++ paper.md
@@ -54,15 +54,15 @@ If the turbulence is isotropic in addition to being homogeneous, the structure f
 In the next section, we provide a brief description of the code.
 
 # Design of the Code
-``Kolmogorov41``  is written in C++. The code uses the Blitz++ library for vectorized array operations. The velocity structure functions ($S_q^{u_\parallel}(l)$ and $S_q^{u_\perp}(l)$) and the scalar structure functions ($S_q^{\theta}(l)$) are computed and stored in a Blitz array. ``Kolmogorov41`` contains separate functions for computing the scalar and the velocity structure functions. Also, separate functions are called for two-dimensional or three-dimensional input fields. However, the basic design of all the functions is the same, as described below.
+``fastSF``  is written in C++. The code uses the Blitz++ library for vectorized array operations. The velocity structure functions ($S_q^{u_\parallel}(\mathbf{l})$ and $S_q^{u_\perp}(\mathbf{l})$) and the scalar structure functions ($S_q^{\theta}(\mathbf{l})$) are computed and stored in a Blitz array. ``fastSF`` contains separate functions for computing the scalar and the velocity structure functions. Also, separate functions are called for two-dimensional or three-dimensional input fields. However, the basic design of all the functions is the same, as described below.
 
-For three-dimensional fields, `Kolmogorov41` employs six nested "for" loops for computation of structure functions using three-dimensional fields. The outer three loops describe the position vector $\mathbf{r}=(x_1,y_1,z_1)$, and the inner three describe the position vector $\mathbf{r+l}=(x_2,y_2,z_2)$. For two-dimensional fields, the code employs four nested "for" loops, with the outer two describing $\mathbf{r}=(x_1,z_1)$ and the inner two describing $\mathbf{r+l}=(x_2,z_2)$. Inside the last loop, the code computes $\delta u_\parallel^q(l)$ and $\delta u_\perp^q(l)$ for the velocity structure functions, or $\delta \theta^q(l)$ for the scalar structure functions, and increments these quantities to the corresponding element of the structure function array. For averaging, the code divides the structure functions array by another array that stores the number of times the innermost loop corresponds to a particular value of $l$.
+Fig. 1 exhibits a brief schematic diagram for the methodology to compute the structure functions. The code navigates through different points in the domain, with each point corresponding to the position vector $\mathbf{l}$. For every $\mathbf{l}$, two subdomains of equal dimensions are considered. The origin of the first subdomain (pink) coincides with that of the main domain. The origin of the second subdomain (green) lies on the position $\mathbf{l}$. The velocity (or scalar) differential field array is then computed by subtracting the velocity (or scalar) field in all the points in the first subdomain from the corresponding points in the second subdomain. The code computes $\delta U_{\parallel}^q$ and $\delta U_{\perp}^q$ arrays for every velocity differential, and then calculates the mean of these arrays to obtain the velocity structure functions. Similarly, the code computes the mean of the scalar differential array to obtain the scalar structure functions. Note that we employ vectorisation in computing the aforementioned arrays as this makes the code efficient and fast. 
 
-Additionally, as per the choice of the user, `Kolmogorov41` can also compute and store the structure functions as functions of the displacement vector $\mathbf{l}$. This form is useful for anisotropic homogeneous turbulence such as rotating or stably-stratified flows.
+Note that an alternative straighter method is to navigate a pair of points through the entire domain and compute the structure functions for every distance between the two points. However, this requires six nested loops for three-dimensional domains (four nested loops for two dimensional domains) because it is not possible to employ vectorisation using this technique. We remark that the strategy used in our code makes it approximately twenty times faster than what it would have been if six nested `for` loops were used for computing the structure functions.
 
[…]
