### 2020-02-19 f1192b3
Update paper.md
