### 2020-02-19 bf654c8
Update paper.md
