+++ paper.md
@@ -86,7 +86,7 @@ Procedure:
         
         * for order $q$:
         
-            * $S_q^{u_{\parallel}} =$ Average of $\delta u_{\parallel}^q$ 
+            * $S_q^{u_{\parallel}} =$ Average of $\delta u_{\parallel}^q$; $S_q^{u_{\perp}} =$ Average of $\delta u_{\perp}^q$ 
   
 
 ```
