+++ paper.md
@@ -86,47 +86,13 @@ Procedure:
         
         * for order $q$:
         
-            * $S_q^{u_{\parallel}} =$ Average of $\delta u_{\parallel}^q$; $S_q^{u_{\perp}} =$ Average of $\delta u_{\perp}^q$ 
-  
-
-```
-    
-for (int ix=starPt; ix<endPt; ix++){
-  int x=index_list(ix, 0, rank_mpi);
-  int z=index_list(ix, 1, rank_mpi);
-  
-  dT.resize(Nx-x,Nz-z);
-  int count=(Nx-x)*(Nz-z);
-  
-  //Calculating the scalar differential
-  dT(Range::all(),Range::all())=T(Range(x,Nx-1),Range(z,Nz-1))
-                                -T(Range(0,Nx-x-1),Range(0,Nz-z-1));
- 
-  for (int p=0; p<=q2-q1; p++){
-    double St = sum(pow(dT(Range::all(),Range::all()),q1+p))/(count);
-    Array<int, 1> X, Z, p_arr;
-    Array<double, 1> St_arr;
-        
-    if (rank_mpi==0) {
-      X.resize(P);
-      Z.resize(P);
-      p_arr.resize(P);
-      St_arr.resize(P);
-    }
-    
-    MPI_Gather(&x, 1, MPI_INT, X.data(), 1, MPI_INT, 0, MPI_COMM_WORLD);
-    MPI_Gather(&z, 1, MPI_INT, Z.data(), 1, MPI_INT, 0, MPI_COMM_WORLD);
-    MPI_Gather(&St, 1, MPI_DOUBLE_PRECISION, St_arr.data(), 
-               1, MPI_DOUBLE_PRECISION, 0, MPI_COMM_WORLD);
-    MPI_Gather(&p, 1, MPI_INT, p_arr.data(), 1, MPI_INT, 0, MPI_COMM_
-    if (rank_mpi==0) {
-        for (int i=0; i<P; i++) {
-            SF_Grid2D_scalar(X(i), Z(i), p_arr(i)) = St_arr(i);
-            } 
-        }
-    }
-}
- ```
+            * $S_q^{u_{\parallel}} =$ Average of $\delta u_{\parallel}^q$; $S_q^{u_{\perp}} =$ Average of $\delta u_{\perp}^q$.
+            
+            * Send the values of $S_q^{u_{\parallel}}$, $S_q^{u_{\perp}}$, $q$, $l_x$, $l_y$, and $l_z$ to the master processor.
+            
+            * If the processor is the master processor, depending on $l_x$, $l_y$, $l_z$, and $q$, store $S_q^{u_{\parallel}}$ and $S_q^{u_{\perp}}$ in the appropriate position in $S_q^{u_{\parallel}}$ and $S_q^{u_{\parallel}}$ arrays respectively.
+            
+* Stop
 
 In the next section, we discuss the validation of our code.
  
