### 2020-02-20 4e43025
Input the number of MPI processors through terminal; Updated the design of code in paper.md
