+++ src/fastSF.cc
@@ -58,7 +58,6 @@ using namespace blitz;
 
 //Function declarations
 void Read_para(); 
-void write_2D(Array<double,2>, string);
 void write_3D(Array<double,3>, string, int);
 void write_4D(Array<double,4>, string, int);
 void read_2D(Array<double,2>, string, string);
@@ -383,6 +382,14 @@ int main(int argc, char *argv[]) {
     MPI_Init(NULL, NULL);
     MPI_Comm_rank(MPI_COMM_WORLD, &rank_mpi);
     MPI_Comm_size(MPI_COMM_WORLD, &P);
+
+    //set the number of processors in x direction
+    if (argc>1) {
+        px = std::atoi(argv[1]);
+    }
+    else {
+        px = 1;
+    }
 
     //Initiallizing h5si
     h5::init();
@@ -1158,25 +1165,6 @@ void compute_time_elapsed(timeval start_t, timeval end_t, double& elapsed){
 }
 
 
-/**
- ********************************************************************************************************************************************
- * \brief   Function to write the structure functions as function of \f$ l \f$ as a 2D hdf5 file.
- *
- *          This function reads writes the structure functions as a 2D array of dimensions \f$ (N_l \times p) \f$. Here, \f$N_l = \sqrt{Nx^2 + Ny^2 + Nz^2}\f$ is the
- *          number of equidistant points from \f$ 0 \f$ to \f$ L \f$, where \f$ L \f$ is the length of the diagonal of the domain in which the structure functions are
- *          calculated. \f$ p \f$ is the order of the structure functions.
- *
- * \param   A is the 2D array to store the structure functions.
- * \param   file is the name of the hdf5 file and the dataset in which the structure functions are stored.
- ********************************************************************************************************************************************
- */
-void write_2D(Array<double,2> A,string file) {
-  int np=A(Range::all(),0).size();
-  int nr=A(0,Range::all()).size();
-  h5::File f("out/"+file+".h5", "w");
-  h5::Dataset ds = f.create_dataset(file, h5::shape(np,nr), "double");
-  ds << A.data();
-}
 
 /**
  ********************************************************************************************************************************************
@@ -1310,9 +1298,6 @@ void Read_para() {
   
     para["structure_function"]["q1"]>>q1;
     para["structure_function"]["q2"]>>q2;
-  
-    para["number_of_proc_along_x"]>>px;
-
     para["test"]["test_switch"]>>test_switch;
   
     if (Nx==1){dx=0;}
@@ -1334,13 +1319,13 @@ void Read_para() {
     }
     if (Nx/2%px != 0) {
         if (rank_mpi==0){
-            cout<<"ERROR! Number of processors in x direction should be less or equal to Nx/4 and some power of 2\n Aborting...\n";
+            cout<<"ERROR! Number of processors in x direction should be less or equal to Nx/2 and some power of 2\n Aborting...\n";
             exit(1);
         }
     }
     if (Ny/2%(P/px) != 0) {
         if (rank_mpi==0){
-            cout<<"ERROR! Number of processors in y direction should be less or equal to Ny/4 and some power of 2\n Aborting...\n";
[…]
