+++ src/Makefile
@@ -40,7 +40,5 @@
  ############################################################################################################################################
 ##
 
-Structure: Kolmogorov41.cc
-	mpic++ Kolmogorov41.cc -O3 -lh5si -lhdf5 -lyaml-cpp -fopenmp -o Kolmogorov41.out
-	#CC main.cc -DBZ_DEBUG -L/home/x_sadhuks/local/lib/libblitz.a -lh5si -lhdf5 -lyaml-cpp -fopenmp -o main.out
-	#CC main.cc -O2 -lh5si -lhdf5 -lyaml-cpp -fopenmp -o main.out
+Structure: fastSF.cc
+	mpic++ fastSF.cc -O3 -lh5si -lhdf5 -lyaml-cpp -o fastSF.out
