+++ paper.md
@@ -58,6 +58,8 @@ In the next section, we provide a brief description of the code.
 
 Fig. 1 exhibits a brief schematic diagram for the methodology to compute the structure functions. The code navigates through different points in the domain, with each point corresponding to the position vector $\mathbf{l}$. For every $\mathbf{l}$, two subdomains of equal dimensions are considered. The origin of the first subdomain (pink) coincides with that of the main domain. The origin of the second subdomain (green) lies on the position $\mathbf{l}$. The velocity (or scalar) differential field array is then computed by subtracting the velocity (or scalar) field in all the points in the first subdomain from the corresponding points in the second subdomain. The code computes $\delta U_{\parallel}^q$ and $\delta U_{\perp}^q$ arrays for every velocity differential, and then calculates the mean of these arrays to obtain the velocity structure functions. Similarly, the code computes the mean of the scalar differential array to obtain the scalar structure functions. Note that we employ vectorisation in computing the aforementioned arrays as this makes the code efficient and fast. 
 
+![A schematic diagram showing the procedure of computing the structure functions using vectorization.\label{Schematic}](Schematic.png)
+
 Note that an alternative straighter method is to navigate a pair of points through the entire domain and compute the structure functions for every distance between the two points. However, this requires six nested loops for three-dimensional domains (four nested loops for two dimensional domains) because it is not possible to employ vectorisation using this technique. We remark that the strategy used in our code makes it approximately twenty times faster than what it would have been if six nested `for` loops were used for computing the structure functions.
 
 We employ Message Passing Interface (MPI) to parallelise the code.  The points ($\mathbf{l}$) of the domain are distributed among the processors along $x$ and $y$ directions for three dimensions, and along $x$ and $z$ directions for two dimensions. Each processor computes the structure functions for the points assigned to it. The input fields are accessible to all the processors to minimize the need of communication. The function `compute_index_list` distributes the load equally among the processors. The master processor stores the structure function array; it receives the structure functions computed for different $\mathbf{l}$ from all the processors and assigns them to the appropriate index of the structure function array.  
