+++ paper.md
@@ -1,5 +1,5 @@
 ---
-title: 'Kolmogorov41: A hybrid parallel code for computing velocity and scalar structure functions'
+title: 'fastSF: A hybrid parallel code for computing velocity and scalar structure functions'
 
 tags:
   - C++
@@ -56,7 +56,7 @@ In the next section, we provide a brief description of the code.
 # Design of the Code
 ``fastSF``  is written in C++. The code uses the Blitz++ library for vectorized array operations. The velocity structure functions ($S_q^{u_\parallel}(\mathbf{l})$ and $S_q^{u_\perp}(\mathbf{l})$) and the scalar structure functions ($S_q^{\theta}(\mathbf{l})$) are computed and stored in a Blitz array. ``fastSF`` contains separate functions for computing the scalar and the velocity structure functions. Also, separate functions are called for two-dimensional or three-dimensional input fields. However, the basic design of all the functions is the same, as described below.
 
-Fig. 1 exhibits a brief schematic diagram for the methodology to compute the structure functions. The code navigates through different points in the domain, with each point corresponding to the position vector $\mathbf{l}$. For every $\mathbf{l}$, two subdomains of equal dimensions are considered. The origin of the first subdomain (pink) coincides with that of the main domain. The origin of the second subdomain (green) lies on the position $\mathbf{l}$. The velocity (or scalar) differential field array is then computed by subtracting the velocity (or scalar) field in all the points in the first subdomain from the corresponding points in the second subdomain. The code computes $\delta U_{\parallel}^q$ and $\delta U_{\perp}^q$ arrays for every velocity differential, and then calculates the mean of these arrays to obtain the velocity structure functions. Similarly, the code computes the mean of the scalar differential array to obtain the scalar structure functions. Note that we employ vectorisation in computing the aforementioned arrays as this makes the code efficient and fast. 
+Fig. 1 exhibits a brief schematic diagram for the methodology to compute the structure functions. The code navigates through different points in the domain, with each point corresponding to the position vector $\mathbf{l}$. $l_x$, $l_y$, and $l_z$ are varied from 0 to $L_x/2$, $L_y/2$, and $L_z/2$ respectively, where $L_x$, $L_y$, and $L_z$ are the dimensions of the domain. Note that structure functions are important for intermediate scales (inertial range) only; thus computation of these quantites for large values of $l$ is not required. For every $\mathbf{l}$, two subdomains of equal dimensions are considered. The origin of the first subdomain (pink) coincides with that of the main domain. The origin of the second subdomain (green) lies on the position $\mathbf{l}$. The velocity (or scalar) differential field array is then computed by subtracting the velocity (or scalar) field in all the points in the first subdomain from the corresponding points in the second subdomain. The code computes $\delta U_{\parallel}^q$ and $\delta U_{\perp}^q$ arrays for every velocity differential, and then calculates the mean of these arrays to obtain the velocity structure functions. Similarly, the code computes the mean of the scalar differential array to obtain the scalar structure functions. Note that we employ vectorisation in computing the aforementioned arrays as this makes the code efficient and fast. 
 
 ![A schematic diagram showing the procedure of computing the structure functions using vectorization.\label{Schematic}](Schematic.png)
 
@@ -101,8 +101,8 @@ In the next section, we discuss the scaling of our code.
  
 # Results
 
-We validate `Kolmogorov41` by using it to compute the structure functions for specific cases, and then comparing our results with those obtained analytically.
[…]
