+++ paper.md
@@ -56,21 +56,17 @@ In the next section, we provide a brief description of the code.
 # Design of the Code
 ``fastSF``  is written in C++. The code uses the Blitz++ library for vectorized array operations. The velocity structure functions ($S_q^{u_\parallel}(\mathbf{l})$ and $S_q^{u_\perp}(\mathbf{l})$) and the scalar structure functions ($S_q^{\theta}(\mathbf{l})$) are computed and stored in a Blitz array. ``fastSF`` contains separate functions for computing the scalar and the velocity structure functions. Also, separate functions are called for two-dimensional or three-dimensional input fields. However, the basic design of all the functions is the same, as described below.
 
-Figure \ref{Schematic} exhibits a schematic diagram for the methodology to compute the structure functions. The code navigates through different points in the domain, with each point corresponding to the position vector $\mathbf{l}=(l_x,l_y,l_z)$.  As shown in Fig. \ref{Schematic}, there are two subdomains of equal dimensions (represented in green and pink colors) for every $\mathbf{l}$. Let us represent the green subdomain $G$ and the pink subdomain by $P$. These subdomains are described as:
-$$G = \{(x,y,z): l_x \leq x \leq L_x, l_y \leq y \leq L_y, l_z \leq z \leq L_z \} $$
-$$P = \{(x,y,z): 0 \leq x \leq L_x-l_x, 0 \leq y \leq L_y-l_y, 0 \leq z \leq L_z-l_z \} $$
-
-The origin of the second subdomain (green) lies on the position $\mathbf{l}$. The velocity (or scalar) differential field array is then computed by subtracting the velocity (or scalar) field in all the points in the first subdomain from the corresponding points in the second subdomain. The code computes $\delta U_{\parallel}^q$ and $\delta U_{\perp}^q$ arrays for every velocity differential, and then calculates the mean of these arrays to obtain the velocity structure functions. Similarly, the code computes the mean of the scalar differential array to obtain the scalar structure functions. Note that we employ vectorisation in computing the aforementioned arrays as this makes the code efficient and fast. 
+Figure \ref{Schematic} exhibits a schematic diagram for the methodology to compute the structure functions. The code navigates through different points in the domain, with each point corresponding to the position vector $\mathbf{l}=(l_x,l_y,l_z)$.  As shown in Fig. \ref{Schematic}, there are two subdomains of equal dimensions (represented in green and pink colors) for every $\mathbf{l}$. Let us represent the green subdomain by $D_{\mathrm{green}}$ and the pink subdomain by $D_{\mathrm{pink}}$. These subdomains are described as:
+$$D_{\mathrm{green}} = \{(x,y,z): l_x \leq x \leq L_x, l_y \leq y \leq L_y, l_z \leq z \leq L_z \} $$
+$$D_{\mathrm{pink}} = \{(x,y,z): 0 \leq x \leq L_x-l_x, 0 \leq y \leq L_y-l_y, 0 \leq z \leq L_z-l_z \} $$
+The velocity (or scalar) differential field array is then computed by subtracting the velocity (or scalar) field in all the points in the subdomain $D_{\mathrm{pink}}$ from the corresponding points in the subdomain $D_{\mathrm{green}}$. The code computes $\delta U_{\parallel}^q$ and $\delta U_{\perp}^q$ arrays for every velocity differential, and then calculates the mean of these arrays to obtain the velocity structure functions. Similarly, the code computes the mean of the scalar differential array to obtain the scalar structure functions. Note that we employ vectorisation in computing the aforementioned arrays. Vectorization makes the code approximately twenty times faster than what it would have been if nested `for` loops were used instead.  
 
 ![A schematic diagram showing the procedure of computing the structure functions using vectorization.\label{Schematic}](Schematic.png)
 
[…]
