+++ paper.md
@@ -64,7 +64,7 @@ The velocity (or scalar) differential field array is then computed by subtractin
 ![A schematic diagram showing the procedure of computing the structure functions using vectorization.\label{Schematic}](Schematic.png)
 
 The current version of ‘fastSF’ computes the structure functions correctly only for homogeneous and isotropic turbulence. To save time and memory, ‘fastSF’ computes and stores the structure functions for only the positive values of $l_x$, $l_y$, and $l_z$. Note that this still gives the correct values for isotropic turbulence because in such case, the structure functions depend only on the magnitude of $\mathbf{l}$ and not its orientation. Further, the structure functions are important for intermediate scales (inertial range) only; thus computation of these quantites for large values of $l$ is not required.
-Therefore, $\mathbf{l}$ is varied from 0 to $L_x/2$, $L_y/2$, and $L_z/2$ respectively, further saving computational resources.
+Therefore, $\mathbf{l}$ is varied up to half the domain size, ($L_x/2, L_y/2, L_z/2$), further saving computational resources.
 
 We employ Message Passing Interface (MPI) to parallelise the code.  The points ($\mathbf{l}$) of the domain are distributed among the processors along $x$ and $y$ directions for three dimensions, and along $x$ and $z$ directions for two dimensions. The distribution of points is done in such a way that every processor gets equal load. Each processor computes the structure functions for the points assigned to it. The input fields are accessible to all the processors to minimize the need of communication. The master processor stores the structure function array; it receives the structure functions computed for different $\mathbf{l}$ from all the processors and assigns them to the appropriate index of the structure function array.  
 
