+++ paper.md
@@ -60,7 +60,7 @@ Figure \ref{Schematic} exhibits a schematic diagram for the methodology to compu
 $$D_{\mathrm{green}} = \{(x,y,z): l_x \leq x \leq L_x, l_y \leq y \leq L_y, l_z \leq z \leq L_z \}, $$
 and the pink subdomain is given by
 $$D_{\mathrm{pink}} = \{(x,y,z): 0 \leq x \leq L_x-l_x, 0 \leq y \leq L_y-l_y, 0 \leq z \leq L_z-l_z \}. $$
-The velocity (or scalar) differential field array is computed by subtracting the velocity (or scalar) field in all the points in the subdomain $D_{\mathrm{pink}}$ from the corresponding points in the subdomain $D_{\mathrm{green}}$. The code computes $\delta U_{\parallel}^q$ and $\delta U_{\perp}^q$ arrays for every velocity differential, and then calculates the mean of these arrays to obtain the velocity structure functions. Similarly, the code computes the mean of the scalar differential array to obtain the scalar structure functions. Note that we employ vectorisation in computing the aforementioned arrays. Vectorization makes the code approximately twenty times faster than what it would have been if nested `for` loops were used instead.  
+The velocity (or scalar) differential field array is computed by vectorized subtractiion of the velocity (or scalar) field of the subdomain $D_{\mathrm{pink}}$ from that of the subdomain $D_{\mathrm{green}}$. The code computes $\delta U_{\parallel}^q$ and $\delta U_{\perp}^q$ arrays for every velocity differential, and then calculates the mean of these arrays to obtain the velocity structure functions. Similarly, the code computes the mean of the scalar differential array to obtain the scalar structure functions. Note that we employ vectorisation in computing the aforementioned arrays. Vectorization makes the code approximately twenty times faster than what it would have been if nested `for` loops were used instead.  
 
 ![A schematic diagram showing the procedure of computing the structure functions using vectorization.\label{Schematic}](Schematic.png)
 
