+++ paper.md
@@ -112,7 +112,6 @@ We perform four runs on a Cray XC40 system (Shaheen II of KAUST) for this proble
  
 # Validation
 
-We validate `fastSF` by using it to compute the structure functions for specific cases, and then comparing our results with those obtained analytically.
 We validate `fastSF` by comparing the numerical results with analytical results for idealized $\mathbf{u}$ and $\theta$ fields as well as with the predictions of K41 [@Kolmogorov:Dissipation; @Kolmogorov:Structure].
 
 ### Problem 1
