### 2020-02-23 a08cf06
Updated documentation. Removed unnecessary functions. Display the no of processors in x and y direction.
