+++ src/fastSF.cc
@@ -1,5 +1,5 @@
 /********************************************************************************************************************************************
- * Kolmogorov41
+ * fastSF
  *
  * Copyright (C) 2020, Mahendra K. Verma
  *
@@ -30,12 +30,12 @@
  ********************************************************************************************************************************************
  */
 
-/*! \file Kolmogorov41.cc
+/*! \file fastSF.cc
  *
  *  \brief Code to compute structure functions using velocity or scalar field data.
  *
- *  \author Shubhadeep Sadhukhan, Shashwat Bhattacharya
- *  \date Jan 2020
+ *  \author Shubhadeep Sadhukhan, Shashwat Bhattacharya, Mahendra K. Verma
+ *  \date Feb 2020
  *  \copyright New BSD License
  *
  ********************************************************************************************************************************************
@@ -57,7 +57,7 @@ using namespace std;
 using namespace blitz;
 
 //Function declarations
-void Read_para(); 
+void get_Inputs(); 
 void write_3D(Array<double,3>, string, int);
 void write_4D(Array<double,4>, string, int);
 void read_2D(Array<double,2>, string, string);
@@ -87,14 +87,12 @@ void Read_Init(Array<double,3>&, Array<double,3>&, Array<double,3>&);
 void Read_Init(Array<double,2>&);
 void Read_Init(Array<double,3>&);
 
-double powInt(double, int);
 
 
 void SF_scalar_3D(Array<double,3>);
 
 
 void SF_scalar_2D(Array<double,2>);
-void compute_index_list(Array<int, 2>&);
 
 void Read_fields();
 void resize_SFs();
@@ -284,10 +282,10 @@ int q2;
 /**
  ********************************************************************************************************************************************
  * \brief   This variable decides whether to calculate both transverse and longitudinal structure functions or only the longitudinal structure
- * functions using a less computationally expensive technique.
+ * functions.
  *
  * If the value is false, the code calculates both transverse and longitudinal structure functions. Else, it calculates only the longitudinal
- * structure functions using less number of iterations.
+ * structure functions.
  ********************************************************************************************************************************************
  */
 bool longitudinal;
@@ -363,17 +361,13 @@ int P;
  ********************************************************************************************************************************************
  */
 int px;
-//$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$//
+
 
 
 /**
  ********************************************************************************************************************************************
- * \brief   The main function of the "Kolmogorov41".
- *
- *          This function is the main function of the "Kolmogorov41" for computing the velocity and scalar structure functions. The MPI
- *          decomposition and integration are also carried out in this function.
- *
- *
+ * \brief   The main function of the "fastSF".
+ 
  ********************************************************************************************************************************************
  */
 int main(int argc, char *argv[]) {
@@ -401,7 +395,7 @@ int main(int argc, char *argv[]) {
     double elapsedt=0.0;
     double elapsepdt=0.0;
 
-    Read_para();
+    get_Inputs();
 
     //Resizing the input fields
     Read_fields();
@@ -451,12 +445,8 @@ int main(int argc, char *argv[]) {
 
 /**
 *************************************************************************************************************************************
-*\brief     Function to ensure that the load is equally distributed among all the MPI processors.
+*\brief     Function to generate or read the input fields.
 *           
-*           This function generates a matrix of different processors with the list of indices to be covered by each processor. The rows of the
-*           matrix denotes the rank of a processor while the columns denote the indices of each processor.
-*
-* \param    index_list stores the aforementioned matrix
 *************************************************************************************************************************************
 */
 void Read_fields() {
@@ -531,12 +521,8 @@ void Read_fields() {
 
 /**
[…]
