+++ README.md
@@ -1,18 +1,18 @@
-# `Kolmogorov41`
+# `fastSF`
 
-`Kolmogorov41` is an open source hybrid parallel C++ code to compute structure functions for a given velocity or scalar field.
+`fastSF` is an open source hybrid parallel C++ code to compute structure functions for a given velocity or scalar field.
 
 ## Getting the Source Code
 
-`Kolmogorov41` is hosted on GitHub. You can download the source code from the following link:
+`fastSF` is hosted on GitHub. You can download the source code from the following link:
 
-https://github.com/ShubhadeepSadhukhan1993/Kolmogorov41
+https://github.com/ShubhadeepSadhukhan1993/fastSF
 
-## Installing `Kolmogorov41`
+## Installing `fastSF`
 
 ### Required Libraries
 
-The following libraries are required for installing and running Kolmogorov41:
+The following libraries are required for installing and running fastSF:
 
 1. `CMake`
 2. `Blitz++`
@@ -27,10 +27,10 @@ The instructions to download and install these libraries are provided in the fol
 
 ###  Compiling instruction
 
-After downloading `Kolmogorov41`, change into `Kolmogorov41/src` directory and run the command `make` in the terminal. An executable named `Kolmogorov41.out` will be created inside the `Kolmogorov41/src` folder.
+After downloading `fastSF`, change into `fastSF/src` directory and run the command `make` in the terminal. An executable named `fastSF.out` will be created inside the `fastSF/src` folder.
 
-## Testing `Kolmogorov41`
-`Kolmogorov41` offers an automated testing process to validate the code. The relevant test scripts can be found in the `tests/` folder of the code. To execute the tesing process, change into `\Kolmogorov41` and run the command 
+## Testing `fastSF`
+`fastSF` offers an automated testing process to validate the code. The relevant test scripts can be found in the `tests/` folder of the code. To execute the tesing process, change into `\fastSF` and run the command 
 
 `bash runTest.sh`. 
 
@@ -40,7 +40,7 @@ The code then runs two test cases; these are as follows.
 
 * In the second case, the code will generate a 2D scalar field given by *T = x + y + z*, and compute the structure functions for the given field. For this case, the structure functions should equal *(l<sub>x</sub> + l<sub>z</sub>)<sup>q</sup>*.
 
-For both the cases, `Kolmogorov41` will compare the computed structure functions with the analytical results. If the percentage difference between the two values is less than 10<sup>-10</sup>, the code is deemed to have passed. 
+For both the cases, `fastSF` will compare the computed structure functions with the analytical results. If the percentage difference between the two values is less than 10<sup>-10</sup>, the code is deemed to have passed. 
 
 Finally, for visualization purpose, the python script `test/test.py` is invoked. This script generates the plots of the second and third-order longitudinal structure functions versus *l*, and the density plots of the computed second-order scalar structure functions and *(l<sub>x</sub> + l<sub>z</sub>)<sup>2</sup>*. These plots demonstrate that the structure functions are computed accurately. Note that the following python modules are needed to run the test script successfully:
 
@@ -49,22 +49,15 @@ Finally, for visualization purpose, the python script `test/test.py` is invoked.
 3. `matplotlib`
 
 
-## Detailed instruction for running `Kolmogorov41`
+## Detailed instruction for running `fastSF`
 
-This section provides a detailed procedure to execute `Kolmogorov41` for a given velocity or scalar field.
+This section provides a detailed procedure to execute `fastSF` for a given velocity or scalar field.
 
[…]
