### 2020-02-23 b0116bb
Minor corrections in the number of processors condition
