+++ in/para.yaml
@@ -5,7 +5,7 @@ program:
     scalar_switch: false
   
     #Please select "true" for 2D operations, "false" for 3D operations:
-    2D_switch : false
+    2D_switch : true
 
     #Please select "true" for computing only the longitudinal structure functions, "false" for computing both the transverse and longitudinal structure functions:
     Only_longitudinal: false
