### 2020-02-23 15601ad
Updated the README.md - Installation of h5si
