### 2020-02-23 10acc72
Updated the test script and the python file
