+++ test/test.py
@@ -1,5 +1,5 @@
 #############################################################################################################################################
- # Kolmogorov41
+ # fastSF
  # 
  # Copyright (C) 2020, Mahendra K. Verma
  #
@@ -94,22 +94,39 @@ X,Y=np.meshgrid(x,y)
 
 
 	
-def plotSF_r2():
-	S_r_Pll2 = hdf5_reader1D("test_velocity_2D/out/SF.h5","SF")[:,1] #
-	S_r_Pll3 = hdf5_reader1D("test_velocity_2D/out/SF.h5","SF")[:,2] #
+def plotSF_r_2D():
+	SF2 = (hdf5_reader_plane("test_velocity_2D/out/SF_Grid_pll2.h5", "SF_Grid_pll2"))
+	SF3 = (hdf5_reader_plane("test_velocity_2D/out/SF_Grid_pll3.h5", "SF_Grid_pll3"))
 	
-	#S_r_Pll_old2 = hdf5_reader1D("out/SF_old2.h5","SF")[:,1] #
-	#S_r_Pll_old3 = hdf5_reader1D("out/SF_old3.h5","SF")[:,1] #
-	r = np.zeros_like(S_r_Pll2) #
-	for i in range(len(S_r_Pll2)): #
-		r[i]=np.sqrt(2)*i/len(S_r_Pll2)     #
+	Nx, Nz = SF2.shape
+	print Nx, Nz
+	Nr = int(np.ceil(np.sqrt(Nx**2 + Nz**2)))
+	print Nr
+
+	r = np.zeros([Nr]) #
+	for i in range(len(r)): #
+		r[i]=np.sqrt(2)*i/(2*len(r))
 	
-	llim = 10
+	SF2_r = np.zeros([Nr])
+	SF3_r = np.zeros([Nr])
+	counter = np.zeros([Nr])
+	
+	for x in range(Nx):
+	    for z in range(Nz):
+	        l = int(np.ceil(np.sqrt(x**2 + z**2)))
+	        SF2_r[l] = SF2_r[l] + SF2[x, z]
+	        SF3_r[l] = SF3_r[l] + SF3[x, z]
+	        counter[l] = counter[l] + 1	  
+	  
+	SF2_r = SF2_r/counter
+	SF3_r = SF3_r/counter
+	
+	llim = 5
 	ulim = 25
 	fig, axes = plt.subplots(figsize = (3.5, 2.57))
 	
-	axes.plot(r[1:len(r)], S_r_Pll2[1:len(r)], color='red', lw=1.5, label=r"$S_2^{u}(l)$") #
-	axes.plot(r[1:len(r)], S_r_Pll3[1:len(r)], color='green', lw=1.5, label=r"$S_3^{u}(l)$")
+	axes.plot(r[1:len(r)], SF2_r[1:len(r)], color='red', lw=1.5, label=r"$S_2^{u}(l)$") #
+	axes.plot(r[1:len(r)], SF3_r[1:len(r)], color='green', lw=1.5, label=r"$S_3^{u}(l)$")
 	
 	
 	s2 = r*r
@@ -125,21 +142,78 @@ def plotSF_r2():
 	axes.set_ylabel('$S_q^{u}(l)$')
 	axes.set_xscale('log')
 	axes.set_yscale('log')
-	axes.set_xlim(0.15, 1)
+	axes.set_xlim(0.1, 0.7)
+	axes.set_ylim(2e-3, 1)
+	axes.set_xticks([0.1, 0.4, 0.7])
+	axes.set_xticklabels([0.1, 0.4, 0.7])
+	axes.legend(scatterpoints=1, loc='lower right', prop={'size':0.95*A}, ncol = 1, frameon=False)
+	fig.tight_layout()
[…]
