+++ cleanTest.sh
@@ -1,7 +1,7 @@
 #!/bin/bash
 
 #############################################################################################################################################
- # Kolmogorov41
+ # fastSF
  # 
  # Copyright (C) 2020, Mahendra K. Verma
  #
