+++ paper.md
@@ -54,7 +54,7 @@ If the turbulence is isotropic in addition to being homogeneous, the structure f
 In the next section, we provide a brief description of the code.
 
 # Design of the Code
-Typical structure function computations in literature involve calculation of $\langle \delta u_\parallel^q \rangle$, $\langle \delta u_\perp^q \rangle$, or $\langle \delta \theta^q \rangle$ using loops over $\mathbf{r}$ and $\mathbf{l}$, which amounts to six nested `for` loops for three dimensions. This makes the computations very expensive for large grids. In our code, we employ vectorization and a single loop over $l$. The new algorithm enhances the performance by a factor of 20 or so over the earlier schemes. Our algorithm for computing the structure functions for three-dimensional velocity field using MPI parallelization is as follows:
+Typical structure function computations in literature involve calculation of $\langle \delta u_\parallel^q \rangle$, $\langle \delta u_\perp^q \rangle$, or $\langle \delta \theta^q \rangle$ using loops over $\mathbf{r}$ and $\mathbf{l}$, which amounts to six nested `for` loops for three dimensions. This makes the computations very expensive for large grids. In our code, we employ vectorization and a single loop over $\mathbf{l}$. The new algorithm enhances the performance by a factor of 20 or so over the earlier schemes. The $\mathbf{l}$'s are divided among MPI processors along $x$ and $y$ directions for three dimensions and along $z$ and $z$ directions for two dimensions. Our algorithm for computing the structure functions for three-dimensional velocity field is as follows:
 
 **Pseudo-code**
 
@@ -62,9 +62,9 @@ Typical structure function computations in literature involve calculation of $\l
 
 *Procedure*:
 
-* For every processor, determine the columns ($l_x, l_y$) assigned to it such that load is evenly distributed among the processors. 
+* For every processor, determine $l_x$ and $l_y$ assigned to it such that load is evenly distributed among the processors. 
 
-* for every column ($l_x, l_y$) assigned to the processor:
+* for every $l_x$ and $l_y$:
      
     * for $l_z from 0 to $L_z$:
         
@@ -85,27 +85,13 @@ Typical structure function computations in literature involve calculation of $\l
              
 * Stop
 
+The calculation procedure is further illustrated in Fig. \ref{Schematic}. The velocity differential $\delta \mathbf{u}(\mathbf{l})$ or the scalar differential $\delta \theta (\mathbf{l})$ is computed by taking a difference between two points with the same index in pink and green subdomains. This feature enables vectorization for computing the velocity or scalar differential. Further, to save computational cost, $\mathbf{l}$ is varied up to half the domain size, ($L_x/2, L_y/2, L_z/2$). Note that the structure functions are important for intermediate scales (inertial range) only; thus, computation of these quantites for large values of $l$ is not required.
 
-``fastSF``  is written in C++. The code uses the Blitz++ library for vectorized array operations. The velocity structure functions ($S_q^{u_\parallel}(\mathbf{l})$ and $S_q^{u_\perp}(\mathbf{l})$) and the scalar structure functions ($S_q^{\theta}(\mathbf{l})$) are computed and stored in a Blitz array. ``fastSF`` contains separate functions for computing the scalar and the velocity structure functions. Also, separate functions are called for two-dimensional or three-dimensional input fields. However, the basic design of all the functions is the same, as described below.
[…]
