+++ paper.md
@@ -54,7 +54,7 @@ If the turbulence is isotropic in addition to being homogeneous, the structure f
 In the next section, we provide a brief description of the code.
 
 # Design of the Code
-Typical structure function computations in literature involve calculation of the velocity or scalar difference using loops over $\mathbf{r}$ and $\mathbf{l}$, which amounts to six nested `for` loops for three dimensions. This makes the computations very expensive for large grids. In our code, we employ vectorization and a single loop over $\mathbf{l}$. The new algorithm enhances the performance by a factor of 20 or so over the earlier schemes. The $\mathbf{l}$'s are divided among MPI processors along $x$ and $y$ directions for three dimensions and along $z$ and $z$ directions for two dimensions. Our algorithm for computing the structure functions for three-dimensional velocity field is as follows:
+Typical structure function computations in literature involve calculation of the velocity or scalar difference using loops over $\mathbf{r}$ and $\mathbf{l}$, which amounts to six nested `for` loops for three dimensions. This makes the computations very expensive for large grids. In our code, we employ vectorization and a single loop over $\mathbf{l}$. The new algorithm enhances the performance by a factor of 20 or so over the earlier schemes. The $\mathbf{l}$'s are divided among MPI processors along $x$ and $y$ directions for three dimensions and along $x$ and $z$ directions for two dimensions. Our algorithm for computing the structure functions for three-dimensional velocity field is as follows:
 
 **Pseudo-code**
 
@@ -97,8 +97,8 @@ We need to compute the structure functions for $l$ ranging from 0 to 7. We divid
 $$\mbox{Processor 1: } l=\{0,7\}, \quad \mbox{Processor 2: } l=\{1, 6\}, $$
 $$\mbox{Processor 3: } l=\{2,5\}, \quad \mbox{Processor 4: } l=\{3, 4\}. $$
 With this distribution, we have the same $\sum l$ ($=7$) for every processor. If two processors are used, then the following distribution results in same $\sum l$ ($=14$) for all the processors.
-$$\mbox{Processor 1: } l=\{0, 7, 1, 6\}, $$
-$$\mbox{Processor 2: } l=\{2, 5, 3, 4\}. $$
+$$\mbox{Processor 1: } l=\{0, 7, 2, 5\}, $$
+$$\mbox{Processor 2: } l=\{1, 6, 3, 4\}. $$
 Thus, each processor is assigned small and large $l$'s (and therefore, large and small loads) to ensure that all the processors get the same total load. This idea has been extended for higher dimensions in our code. However, the algorithm for distribution of load for higher dimensions is complex and the reader can refer to the code for details.
 
 Finally, we remark that the current version of ‘fastSF’ computes the structure functions correctly only for homogeneous and isotropic turbulence. This is because ‘fastSF’ computes and stores the structure functions for only the positive values of $l_x$, $l_y$, and $l_z$. Note that this still gives the correct values for isotropic turbulence because in such case, the structure functions depend only on the magnitude of $\mathbf{l}$ and not its orientation.
