### 2020-02-26 40f72c8
Update paper.md
