### 2020-02-27 b9b6cd4
Changed MPI_DOUBLE_PRECISION to MPI_DOUBLE
