### 2020-02-28 0e9bec2
Updated load distribution in paper.md
