### 2020-02-28 88f71ae
Updated load distribution in paper 1.md
