### 2020-02-28 ed8040c
Updated load distribution in paper 2.md
