+++ paper.md
@@ -89,9 +89,9 @@ The calculation procedure is further illustrated in Fig. \ref{Schematic}. In our
 
 The $\mathbf{l}$'s are divided among MPI processors along $x$ and $y$ directions for three dimensions and along $x$ and $z$ directions for two dimensions. Each MPI processor computes the structure functions for the points assigned to it and has access to the entire input data. Thus, we save considerable time that would otherwise be spent on communication between the processors during the calculation of the velocity or the scalar difference. After computing the structure function for a given $\mathbf{l}$, each processor communicates the result to the master processor, which stores the $S_q^{u_\parallel}(\mathbf{l})$, $S_q^{u_\perp}(\mathbf{l})$ and $S_q^{\theta}(\mathbf{l})$ arrays.
 
-It should be noted that the size of the pink or the green subdomain, which is given by $(L_x-l_x)(L_y-l_y)(L_z-l_z)$, is not the same for different $\mathbf{l}$'s, rather, it decreases with increasing $\mathbf{l}$. Thus, to compute the structure functions, there will be more load for small $\mathbf{l}$'s and less load for large $\mathbf{l}$'s. Because of this, a straightforward division of the domain among the processors along $x$ and $y$ directions will lead to load imbalance. In order to divide the load equally among the processors, $\mathbf{l}'s$ must be distributed in such a way that $\sum \{(L_x-l_x)(L_y-l_y)(L_z-l_z)\}$ is the same for every processor. Therefore, we assign both large and small $\mathbf{l}$'s to each processor to achieve equal load distribution. We illustrate the idea of dividing the load in the following.
+It should be noted that the size of the pink or the green subdomain, which is given by $(L_x-l_x)(L_y-l_y)(L_z-l_z)$, is not the same for different $\mathbf{l}$'s, rather, it decreases with increasing $\mathbf{l}$. Thus, to compute the structure functions, there will be more load for small $\mathbf{l}$'s and less load for large $\mathbf{l}$'s. Because of this, a straightforward division of the domain among the processors along $x$ and $y$ directions will lead to load imbalance. In order to divide the load equally among the processors, $\mathbf{l}$'s must be distributed in such a way that $\sum \{(L_x-l_x)(L_y-l_y)(L_z-l_z)\}$ is the same for every processor. Therefore, we assign both large and small $\mathbf{l}$'s to each processor to achieve equal load distribution. We illustrate the idea of dividing the load in the following.
 
-Consider a one-dimensional domain given by
+Consider a one-dimensional domain of size $L=15$, given by
 $$l=\{0, 1, 2, 3 ... 15\}.$$ 
 We need to compute the structure functions for $l$ ranging from 0 to 7. We divide the task among four processors, with 2 points assigned to each processor. The following distribution of points ensures equal load distribution:
 $$\mbox{Processor 1: } l=\{0,7\}, \quad \mbox{Processor 2: } l=\{1, 6\}, $$
