### 2020-02-28 d99c661
Updated load distribution in paper 3.md
