+++ paper.md
@@ -95,10 +95,7 @@ It is clear from Fig. \ref{Schematic} that the size of the pink or the green sub
 Consider a one-dimensional domain of size $L=15$; possible $l$'s are
 $$l=\{0, 1, 2, 3 ... 15\}.$$ 
 We need to compute the structure functions for $l$ ranging from 0 to 7. We divide the task among four processors, with 2 points assigned to each processor. The following distribution of points ensures equal load distribution:
-$$\mbox{Processor 0: } \quad l=\{0,7\}, \quad \sum(L-l)=(15-0)+(15-7) = 23,$$
-$$\mbox{Processor 1: } \quad l=\{1, 6\}, \quad \sum(L-l)=(15-1)+(15-6) = 23,$$
-$$\mbox{Processor 2: } \quad l=\{2,5\}, \quad \sum(L-l)=(15-2)+(15-5) = 23,$$
-\mbox{Processor 3: } \quad l=\{3, 4\}, \quad \sum(L-3)=(15-1)+(15-4) = 23.$$
+
 Similarly, if two processors are used, then the following distribution results in perfect load balance. 
 $$\mbox{Processor 1: } \quad l=\{0, 7, 2, 5\},$$
 $$\mbox{Processor 2: } \quad l=\{1, 6, 3, 4\}.$$
