+++ paper.md
@@ -112,7 +112,7 @@ $$\mbox{Processor 3: } \quad l=\{3, 4\}, \quad \sum(L-l)=(15-3)+(15-4) = 23.$$
 Similarly, if two processors are used, then the following distribution results in load balance. 
 $$\mbox{Processor 0: } \quad l=\{0, 7, 2, 5\},$$
 $$\mbox{Processor 1: } \quad l=\{1, 6, 3, 4\}.$$
- This idea of load distribution has been implemented in our program and has been extended for higher dimensions. 
+ This idea of load distribution has been implemented in our program and has been extended to higher dimensions. 
 
 Note that for 2D, $l_x>0$, but $l_z$ can take both positive and negative values. However, for isotropic turbulence, the structure functions for $+l_z$ and $-l_z$ are statistically equal. Therefore, in our computations, we restrict  to $l_x>0$, $l_z>0$. For anisotropic turbulence, not discussed here, the structure functions will depend on $(l_x,l_z)$ rather than $l$; our code will be extended to such systems in future. 
 
