### 2020-03-02 f01da8e
Installation instructions of the libraries in README.md
