### 2020-03-02 d8b5720
Delete test.py
