### 2020-03-02 #2086 @shashwat-bhattacharya
Hello @jedbrown
We thank you for your valuable suggestions to improve the code. We have made major changes to the code to make it faster and more efficient. The following are the changes.

1) As per your suggestion, we have implemented vectorization and reduced the number of nested loops in our code.  The new scheme uses only three nested loops instead of six for 3D grids and two nested loops instead of four for 2D grids.  This enhances the performance of our code by a factor of 20 times or so compared to the earlier scheme. Please refer to the paper for a detailed description of the new algorithm. We believe that this algorithm is a value-added feature because typical structure function computations in literature employ schemes similar to the one used in our earlier code.

2) In the current algorithm, a straightforward distribution of tasks among the processors along x and y directions will result in a load imbalance. Therefore, we implement an algorithm to distribute the computational load equally among the processors. Please refer to the paper and the code for details of the load-distribution scheme. Our code is scalable over many processors due to equal load distribution.

3)  The earlier code employed a combination of OpenMP and MPI parallelization for computing the structure functions. However, the algorithm for equal load distribution discussed above is very complex to implement using OpenMP. Therefore, we have removed OpenMP in the new code and employ MPI parallelization along both x and y (or z) directions.

4) In the new code, the computed structure function arrays are stored only in the master node. However, if the input fields are stored in one processor, we will require a lot of inter-processor communication for computing the velocity or scalar difference.  Thus, all the processors read the input data in our code, saving communication time. We believe that sharing the input data in all the processors but storing the output in one processor is the most optimal configuration for our case.

5) The library “H5SI” now contains a BSD license and is available in Github.  The link is given below:
https://github.com/anandogc/h5si
Our code currently uses this version of H5SI.

6) Finally, we have renamed our code to “fastSF”. Further, the title of our paper has been modified to “fastSF: A parallel code for computing the structure functions of turbulence”.

We thank you again for your constructive comments and hope that the code will be widely used in the turbulence community.
