### 2020-03-04 #2086 @shashwat-bhattacharya
@danielskatz We have made changes to the code as per the suggestions of @jedbrown .  A summary of these changes is given in the previous comment. Please note that the name of the code has changed to "fastSF", and the modified title is "fastSF: A parallel code for computing the structure functions of turbulence". The modified code and the paper have been pushed to the GitHub repository.
Could you please restart the review process?
Thank you.
