### 2020-03-04 #2086 @danielskatz
👋 @jedbrown - thanks for considering this - I asked you because of the author's initial request for you, and your involvement so far, but if you feel like you are overloaded with other JOSS submissions, we can move it to someone else. Any suggestions you have for the next best editor would be welcome in this case.
