### 2020-03-04 #2086 @jedbrown
@openjournals/dev :wave: While you're here, we'll also need the repository URL updated. (I edited in the comment, but that doesn't show up in Whedon's PDF.)

@danielskatz I always enjoy seeing @labarba at fluid dynamics workshops :-).  I can take this so long as folks can accept a bit slower responses from me.
