### 2020-03-04 #2086 @arfon
> @openjournals/dev 👋 While you're here, we'll also need the repository URL updated. (I edited in the comment, but that doesn't show up in Whedon's PDF.)

Sorry for the slow response. I somehow managed to unsubscribe from the @openjournals/dev notifications.

I've made both of the requested changes now to the author GitHub handle and the repository URL.
