### 2020-05-26 #2185 @iljah
hi @ShubhadeepSadhukhan1993 , above article proof seems to contain things that aren't supposed to be there according to review guidelines ["the JOSS paper (the compiled PDF associated with this submission) should only include"] and hence should probably be removed before acceptance. @jedbrown probably knows better but I'd say the following sections should be removed: "Velocity and scalar structure functions", "Design of the Code", "Validation".
