### 2020-05-26 bfb76ab
Update paper.md according to referee's suggestion.
Removed three sections.
