+++ paper.md
@@ -47,82 +47,15 @@ Turbulence is a complex phenomenon in fluid dynamics involving nonlinear interac
 6.   cmake (version 3.16.4) 
 
 
-In the next section, we will briefly define the velocity and the scalar structure functions in turbulence.
+In the next section, we will briefly discuss the usefulness of the code to the turbulence community.
 
-# Velocity and scalar structure functions
 
-We denote the velocity and scalar fields using $\mathbf{u}$ and $\theta$  respectively. The velocity difference between any two points $\mathbf{r}$ and $\mathbf{r+l}$ is $\delta \mathbf{u} = \mathbf{u(r+l)}-\mathbf{u(r)}$. The difference in the parallel components of the velocity field along $\mathbf{l}$ is $\delta u_\parallel=\delta \mathbf{u}\cdot \hat{\mathbf{l}}$.  The corresponding difference in the perpendicular component is $\delta u_\perp= |\delta \mathbf{u} - \delta u_\parallel \hat{\mathbf{l}}|$. Assuming statistical homogeneity, we define the longitudinal velocity structure functions of order $q$ as
-$$ S_q^{u_\parallel}(\mathbf{l}) = \langle (\delta u_\parallel)^q \rangle = \langle [\{\mathbf{u(r+l)}-\mathbf{u(r)}\}\cdot \hat{\mathbf{l}}]^q \rangle, \quad \quad (1)$$ 
-and the transverse velocity structure functions of order 
-$q$ as 
-$$ S_q^{u_\perp}(\mathbf{l}) = \langle (\delta u_\perp)^q \rangle = \langle |\delta \mathbf{u} - \delta u_\parallel \hat{\mathbf{l}}|^q \rangle. \quad \quad (2)$$ 
-Here, $\langle \cdot \rangle$ denotes ensemble averaging. Similarly, we can define the scalar structure functions for the scalar field as 
-$$ S_q^\theta(\mathbf{l}) = \langle (\delta \theta)^q\rangle = \langle [\theta (\mathbf{r+l}) - \theta(\mathbf{r})]^q \rangle. \quad \quad(3)$$
 
-For isotropic turbulence (in addition to being homogeneous), the structure functions become functions of $l$, where $l=|\mathbf{l}|$. The second-order velocity structure function $S_q^{u_{\parallel}}(l)$ provides an estimate for the energy in the eddies of size $l$ or less [@Davidson:book:Turbulence]. 
+# Performance and Scaling of `fastSF`
 
-In the next section, we provide a brief description of the code.
+Typical structure function computations in literature involve calculation of the velocity or scalar difference using loops over two points. These computations require six nested `for` loops for 3D fields that makes the computations very expensive for large grids. `fastSF` is scalable over many processors due to vectorization and equal load distribution. In our code, we employ vectorization and loops over only one point, thus requiring three loops instead of six for 3D fields. The new algorithm enhances the performance approximately 20 times over the earlier schemes due to vectorization. We further ensure equal load distribution among processors to enhance scalabality of the code over many processors. Please refer to the code and the documentation for details.
 
-# Design of the Code
[…]
