+++ paper.md
@@ -53,7 +53,7 @@ In the next section, we will briefly discuss the performance and scaling of ``fa
 
 # Performance and scaling of `fastSF`
 
-Typical structure function computations in literature involve calculation of the velocity or scalar difference using loops over two points. These computations require six nested `for` loops for 3D fields that makes the computations very expensive for large grids. `fastSF` is scalable over many processors due to vectorization and equal load distribution. In our code, we employ vectorization and loops over only one point, thus requiring three loops instead of six for 3D fields. The new algorithm enhances the performance approximately 20 times over the earlier schemes due to vectorization. We further ensure equal load distribution among processors to enhance scalabality of the code over many processors. Please refer to the code and the documentation for details.
+Typical structure function computations in literature involve calculation of the velocity or scalar difference using loops over two points. These computations require six nested `for` loops for 3D fields that makes the computations very expensive for large grids. `fastSF` is scalable over many processors due to vectorization and equal load distribution. In our code, we employ vectorization and loops over only one point, thus requiring three loops instead of six for 3D fields. The new algorithm enhances the performance approximately 20 times over the earlier schemes due to vectorization. Further, we ensure equal load distribution among processors to enhance scalabality of the code over many processors. Please refer to the code and the documentation for details.
 
 We demonstrate the scaling of `fastSF` for the third-order longitudinal structure function for an idealized velocity field on a $128^3$ grid.  For our computation we employ a maximum of 1024 processors. We take the velocity field as
 $$\mathbf{u} = 
