### 2020-05-26 #2185 @jedbrown
It's your discretion, but defining what structure functions are and the basic parallel distribution strategy is reasonable to retain (it helps a reader understand what the software really does). Validation (seems more like "verification") could be deferred to project documentation. The important point is to write for the correct audience (first impression on a potential user).
