### 2020-05-26 #2185 @ShubhadeepSadhukhan1993
@iljah Thanks for your view on our paper. Please mention should we retain the paper content as it is or remove the sections as mentioned earlier. We will work on them accordingly. Presently, we have removed those sections from our repository already.
