### 2020-06-06 #2185 @jedbrown
It's your discretion based on what you think a reader will need to know to assess whether your package is likely to meet their needs.
