### 2020-06-07 3a2d2c4
Changed paper.md and added Verification.md
