### 2020-06-07 331f0c2
Update Verification.md
math part in verification.md is updated
