+++ Verification.md
@@ -1,41 +1,41 @@
 # Testing and Validation
 
-We validate `fastSF` by comparing the numerical results with analytical results for idealized $\mathbf{u}$ and $\theta$ fields as well as with the predictions of K41 [@Kolmogorov:Dissipation; @Kolmogorov:Structure].
+We validate `fastSF` by comparing the numerical results with analytical results for idealized \f$\mathbf{u}\f$ and \f$\theta\f$ fields as well as with the predictions of K41 [@Kolmogorov:Dissipation; @Kolmogorov:Structure].
 
 ### Problem 1
 
 We consider the following 2D velocity and scalar fields:
-$$\mathbf{u} = 
+\f$\mathbf{u} = 
 \begin{bmatrix} 
 x \\ z
 \end{bmatrix}, \quad \theta = x+z.
-$$
+\f$
 For the above fields, it can be analytically shown that the longitudinal and the transverse velocity structure functions and the scalar structure functions are
-$$S_q^{u_\parallel} = (l_x^2 + l_z^2)^{q/2} = l^q,$$
-$$S_q^{u_\perp} = 0,$$
-$$S_q^\theta = (l_x+l_z)^q.$$
-We run ``fastSF`` to compute the velocity and scalar structure functions for the above fields. The resolution of the fields and the domain size are $32^2$ and $1 \times 1$ respectively. We plot the second and the third-order longitudinal velocity structure functions versus $l$ in Fig. \ref{SFTest}. Clearly, $S_2^{u_\parallel}(l)$ and $S_3^{u_\parallel}(l)$ equal $l^2$ and $l^3$ respectively, consistent with the analytical results. Figure \ref{SFScalar} exhibits the density plots of the computed second-order scalar structure function $S_2^{\theta}(\mathbf{l})$ along with $(l_x + l_z)^2$. The two plots are very similar, thus showing the ``fastSF`` computes the scalar structure function correctly.
+\f$S_q^{u_\parallel} = (l_x^2 + l_z^2)^{q/2} = l^q,\f$
+\f$S_q^{u_\perp} = 0,\f$
+\f$S_q^\theta = (l_x+l_z)^q.\f$
+We run ``fastSF`` to compute the velocity and scalar structure functions for the above fields. The resolution of the fields and the domain size are \f$32^2\f$ and \f$1 \times 1\f$ respectively. We plot the second and the third-order longitudinal velocity structure functions versus \f$l\f$ in Fig. \ref{SFTest}. Clearly, \f$S_2^{u_\parallel}(l)\f$ and \f$S_3^{u_\parallel}(l)\f$ equal \f$l^2\f$ and \f$l^3\f$ respectively, consistent with the analytical results. Figure \ref{SFScalar} exhibits the density plots of the computed second-order scalar structure function \f$S_2^{\theta}(\mathbf{l})\f$ along with \f$(l_x + l_z)^2\f$. The two plots are very similar, thus showing the ``fastSF`` computes the scalar structure function correctly.
 
-![For the velocity field defined in Problem 1: plots of the second and third-order longitudinal structure functions vs. $l$. The second and third-order structure functions equal $l^2$ and $l^3$ respectively.\label{SFTest}](SF_test.png)
+![For the velocity field defined in Problem 1: plots of the second and third-order longitudinal structure functions vs. \f$l\f$. The second and third-order structure functions equal \f$l^2\f$ and \f$l^3\f$ respectively.\label{SFTest}](SF_test.png)
 
[…]
