### 2020-06-08 #2185 @shashwat-bhattacharya
@jedbrown Thanks for your reply. We will go ahead with your suggestion of retaining the definition of structure functions and basic parallelization strategy. We will move the validation part to the project documentation. We are currently making these changes and will update you when done.
