### 2020-06-11 #2185 @iljah
Having trouble installing one dependency: https://github.com/anandogc/h5si/issues/1
