### 2020-06-11 #2185 @shashwat-bhattacharya
@iljah Thanks for pointing out the issue. I just had a conversation with the developer of h5si. He will look into it and address the problem in the github page.
From what I have seen in the link shared by you, the error message suggests that your c++ (CXX) compiler does not have mpi. Is it possible that you have mpi installed in your c compiler but not c++?
Anyway, the h5si developer should be able to give you a better advice.
Thanks.
