### 2020-06-11 #2185 @iljah
I have mpicxx and mpicc but h5si doesn't seem to be using them.
