### 2020-06-29 #2185 @iljah
Found another issue with prerequisite h5si: https://github.com/anandogc/h5si/issues/2
