### 2020-07-01 ShubhadeepSadhukhan1993/fastSF#1 @cpgr
Suggested improvements for JOSS publication
Dear authors,

As per the review underway in https://github.com/openjournals/joss-reviews/issues/2185, I suggest the following minor improvements

- Some guidelines for contributing/reporting or suggestions bugs (one of the items to be checked off in the review). This need only be a small section in `README.md` so would not be too onerous
- Add a link to the verification documentation that was removed from the original manuscript. At the moment, I don't think it is linked in the documentation.
- The figures and verification documentation could be moved to the `docs` folder to make the main repository cleaner
- There are some backup files (eg `cleanTests.sh~`) in the repository that should be removed
- Perhaps make it clearer when running the automated test script that the tests passed. I had to look hard to find out the `TEST_PASSED` lines. Perhaps a summary at the end would make it easier for users.
- I get a `H5pyDeprecationWarning` when running the tests. Line 76 in `test.py`
- The documentation notes a sample input field in a `hdf5` file in the `in` directory, but there is no such file that I can find. Perhaps an example using such an input field could be included?
- I also had difficulty installing all of the dependencies on a clean ubuntu box.  It certainly needs more detailed installation instructions. Perhaps you could attempt to install it following your instructions on a fresh install to see if you have any issues as well. In my experience, if a code is difficult to install first time, a lot of potential users will not bother proceeding and will look for alternatives.
- I couldn't really tell straight away what `h5si` does or why I needed it. Perhaps a note to explain it would be useful?
- The links for `HDF5` and `Yaml-cpp` go to a website for `Tarang`

All of these are relatively minor, so shouldn't be too hard to address if you choose.
