### 2020-07-01 #2185 @iljah
@ShubhadeepSadhukhan1993 I don't see any community guidelines in README as required by checklist: Community guidelines: Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support
