### 2020-07-01 #2185 @iljah
@ShubhadeepSadhukhan1993 Could you also please add bit of info to paper about other similar packages, as required by checklist? (State of the field: Do the authors describe how this software compares to other commonly-used packages?)
