### 2020-07-01 #2185 @iljah
And changes I suggested in https://github.com/openjournals/joss-reviews/issues/2185#issuecomment-641971304 are still relevant. At least processors should be changed to cores.
