### 2020-07-01 ShubhadeepSadhukhan1993/fastSF#2 @iljah
3D test?
It would be easier to test scalability of fastSF in realistic scenario if also 3D test program was included. Currently looks like included test programs ignore y dimension.
