### 2020-07-01 #2185 @iljah
@jedbrown does performance evaluation also include scalability? Test programs included in repository are only 2D so it might be difficult to evaluate scalability claims made in article. I suggested adding 3D test program in https://github.com/ShubhadeepSadhukhan1993/fastSF/issues/2
