### 2020-07-01 #2185 @iljah
Based on 2D tests on my laptop the stated performance seems to check out in that it would take on the order of 1000 s to calculate a 1024^2 case if my laptop had 16 cores.
