### 2020-07-01 #2185 @jedbrown
A 3D test problem seems appropriate. Performance evaluation is whatever is claimed. So if the package makes claims about scalability, it'd be reasonable to attempt to confirm that the claims are plausible (within reason; e.g., use what hardware you have handy and that two points on a scalability curve are roughly comparable to the claims). This is mostly about making sure that someone with time/resources could run a reproduction study and as a sanity check.
