### 2020-07-02 dcaffa7
Update README.md
Added community guidelines
