+++ test/test.py
@@ -79,7 +79,11 @@ def hdf5_reader_plane(filename,dataset):
 	return V1
 
 
-
+def hdf5_reader_slice(filename,dataset, Ny):
+	file_V1_read = h5py.File(filename)
+	dataset_V1_read = file_V1_read["/"+dataset]
+	V1=dataset_V1_read[:,Ny,:]
+	return V1
 A=9.3
 font = {'family' : 'serif', 'weight' : 'normal', 'size' : A}
 plt.rc('font', **font)
@@ -92,8 +96,6 @@ y=np.linspace(0,1,N)
 X,Y=np.meshgrid(x,y)
 
 
-#print T[90,100,500]
-#print T[:,255,255]
 
 
 	
@@ -103,8 +105,8 @@ def plotSF_r_2D():
 	
 	Nx, Nz = SF2.shape
 	
-	Nr = int(np.ceil(np.sqrt(Nx**2 + Nz**2)))
-	
+	#Nr = int(np.ceil(np.sqrt(Nx**2 + Nz**2)))
+	Nr = int(np.ceil(np.sqrt((Nx-1)**2 + (Nz-1)**2)))+1
 
 	r = np.zeros([Nr]) #
 	for i in range(len(r)): #
@@ -151,17 +153,18 @@ def plotSF_r_2D():
 	axes.set_xticklabels([0.1, 0.4, 0.7])
 	axes.legend(scatterpoints=1, loc='lower right', prop={'size':0.95*A}, ncol = 1, frameon=False)
 	fig.tight_layout()
-	plt.savefig("SF_test.png", dpi=600)
+	plt.savefig("SF_velocity_r2D.png", dpi=600)
 	plt.show()
 
 
 def plotSF_r_3D():
-	SF2 = (hdf5_reader("out/SF_Grid_pll2.h5", "SF_Grid_pll2"))
-	SF3 = (hdf5_reader("out/SF_Grid_pll3.h5", "SF_Grid_pll3"))
+	SF2 = (hdf5_reader("test_velocity_3D/out/SF_Grid_pll2.h5", "SF_Grid_pll2"))
+	SF3 = (hdf5_reader("test_velocity_3D/out/SF_Grid_pll3.h5", "SF_Grid_pll3"))
 	
 	Nx, Ny, Nz = SF2.shape
 	
-	Nr = int(np.ceil(np.sqrt(Nx**2 + Ny**2 + Nz**2)))
+	#Nr = int(np.ceil(np.sqrt(Nx**2 + Ny**2 + Nz**2)))
+	Nr = int(np.ceil(np.sqrt((Nx-1)**2 + (Ny-1)**2 + (Nz-1)**2)))+1
 	
 
 	r = np.zeros([Nr]) #
@@ -179,7 +182,7 @@ def plotSF_r_3D():
 	           SF2_r[l] = SF2_r[l] + SF2[x, y, z]
 	           SF3_r[l] = SF3_r[l] + SF3[x, y, z]
 	           counter[l] = counter[l] + 1	  
-	  
+	
 	SF2_r = SF2_r/counter
 	SF3_r = SF3_r/counter
 	
@@ -210,20 +213,15 @@ def plotSF_r_3D():
 	axes.set_xticklabels([0.2, 0.4, 0.8])
 	axes.legend(scatterpoints=1, loc='lower right', prop={'size':0.95*A}, ncol = 1, frameon=False)
 	fig.tight_layout()
-	plt.savefig("SF_test.png", dpi=600)
+	plt.savefig("SF_velocity_r3D.png", dpi=600)
 	plt.show()	
 
 	
-def plot_SF_density():
[…]
