### 2020-07-05 cad8f4c
Update paper.md
Added few lines about other packages
