### 2020-07-05 42b1cc8
Update paper.md
Processors changed to cores; cores per node added.
