### 2020-07-05 5dc2eec
Moved the figures to docs
