### 2020-07-10 cf1131c
Delete Verification.md
