+++ docs/Verification.md
@@ -1,0 +1,109 @@
+# Testing and Validation of `fastSF`
+
+We validate `fastSF` by comparing the numerical results with analytical results for idealized **u** and *&theta;* fields as well as with the predictions of K41 (Kolmogorov 1941a, 1941b).
+
+### Problem 1
+
+We consider the following 2D velocity and scalar fields: 
+
+**u** = [*x, z*]
+
+*&theta; = x + z*
+
+For the above fields, it can be analytically shown that the longitudinal velocity structure functions and the scalar structure functions are
+
+*S<sub>q</sub><sup>u<sub>ll</sub></sup> = (l<sub>x</sub><sup>2</sup> + l<sub>z</sub><sup>2</sup>)<sup>q/2</sup> = l<sup>q</sup>*,
+
+*S<sub>q</sub><sup>&theta;</sup> = (l<sub>x</sub> + l<sub>z</sub>)<sup>q</sup>*.
+
+We run ``fastSF`` to compute the velocity and scalar structure functions for the above fields. The resolution of the fields and the domain size are 32<sup>2</sup> and 1 x 1 respectively. We plot the second and the third-order longitudinal velocity structure functions versus *l* in Fig.1. Clearly, *S<sub>2</sub><sup>u<sub>ll</sub></sup> (l)* and *S<sub>3</sub><sup>u<sub>ll</sub></sup> (l)* equal *l<sup>2</sup>* and *l<sup>3</sup>* respectively, consistent with the analytical results. Figure 2 exhibits the density plots of the computed second-order scalar structure function *S<sub>2</sub><sup>&theta;</sup>* (**l**) along with (*l<sub>x</sub> + l<sub>z</sub>*)<sup>2</sup>. The two plots are very similar, thus showing the ``fastSF`` computes the scalar structure function correctly.
+
+
+ <figure>
+  <img src="docs/figs/SF_velocity_r2D.png" alt="Trulli" style="width:70%">
+  <figcaption>Fig.1: For the velocity field defined in Problem 1: plots of the second and third-order longitudinal structure functions vs. l. The second and third-order structure functions equal l<sup>2</sup> and l<sup>3</sup> respectively.</figcaption>
+</figure>
+ 
+<figure>
+  <img src="docs/figs/SF_scalar2D.png" alt="Trulli" style="width:100%">
+  <figcaption>Fig.2: For the scalar field defined in Problem 1: (a) Density plot of the second-order scalar structure function as function of the displacement vector. (b) Density plot of (l<sub>x</sub> + l<sub>z</sub>)<sup>2</sup>, which is the analytical value of the second-order scalar structure function. The two density plots are very similar.</figcaption>
+</figure>
+
+
+### Problem 2
+We consider the following 3D velocity and scalar fields: 
+
+**u** = [*x, y, z*]
+
+*&theta; = x + y + z*
+
+For the above fields, it can be analytically shown that the longitudinal velocity structure functions and the scalar structure functions are
+
+*S<sub>q</sub><sup>u<sub>ll</sub></sup> = (l<sub>x</sub><sup>2</sup> + l<sub>y</sub><sup>2</sup> + l<sub>z</sub><sup>2</sup>)<sup>q/2</sup> = l<sup>q</sup>*,
+
+*S<sub>q</sub><sup>&theta;</sup> = (l<sub>x</sub> + l<sub>y</sub> + l<sub>z</sub>)<sup>q</sup>*.
+
[…]
