### 2020-07-10 d426134
Delete Verification.md
