### 2020-07-10 ShubhadeepSadhukhan1993/fastSF#1 @shashwat-bhattacharya
@cpgr Thanks a lot for your suggestions. We have incorporated them as follows.

* We have added a section "Contributions and bug reports" towards the end of the `README.md` file, in which we briefly explain to contribute and report bugs / issues, if any.

* We have added the link to the verification documentation in the section "Documentation and validation" in the `README.md` file.

* The Verification.md file has been moved to `docs` folder. The figures have been moved to `doc/figures`.

* All the backup files have been removed.

* We agree that the messages stating that the tests have passed were not very conspicuous. We have added the statement `The structure functions computed numerically using the code match with the analytically obtained values.` right after the message`TEST_PASSED`. We believe this should make it easier for the users to spot the message.

* Perhaps the deprication warning was coming because we had not specified how the python script accesses the `hdf5` output files (that is, as read-only, or with write access, etc.). We believe that we are using an older version of `h5py` and therefore not getting the above warning. In the revised script, we specify that the script accesses the `hdf5` file as read-only; hopefully this will resolve the issue. Please let us know if the warning persists.

* We have included an example input field in `hdf5` format in the `in` directory.

* We have now included detailed installation instructions for each library in the `README.md` file. We hope the users do not face difficulty in installing the files. Please let us know in case of any further issue regarding the installation of the libraries.

* In the revised readme file, we briefly describe the functions of each library. `h5si` is a C++ library that basically makes the  reading or writing the`hdf5` files much simpler for the programmer. Note that using `h5si` library, one needs to write only 3 to 4 lines of code for reading or writing `hdf5` files, as opposed to writing many lines of code when only `hdf5` library is used.  We will request the developer of `h5si` to write about the usefullness of this library in its Github page.

* The links for `Yaml-cpp` now redirects the user to the Github page of `yaml-cpp` (version 0.3.0).  Clicking on the link for `HDF5` downloads  `hdf5` (version 1.8.20).

Thanks again for your helpful comments. I hope the issues are resolved. If you have more queries or suggestions regarding the code, we will be happy to work on them.
