### 2020-07-10 #2185 @shashwat-bhattacharya
> @ShubhadeepSadhukhan1993 I don't see any community guidelines in README as required by checklist: Community guidelines: Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support

Thanks @iljah for the suggestion. We have added a section "Contributions and bug reports" towards the end of the README.md file, in which we briefly explain to contribute and report bugs / issues, if any.
