### 2020-07-10 #2185 @shashwat-bhattacharya
> @ShubhadeepSadhukhan1993 Could you also please add bit of info to paper about other similar packages, as required by checklist? (State of the field: Do the authors describe how this software compares to other commonly-used packages?)

@iljah We thank you for the suggestion. To the best of our knowledge and belief, no other commercial or open-source packages for computing structure functions exist or are available at the moment. Currently, such codes are developed in-house and are not open to the general public. We have added the above statements in the section "Summary" of the revised paper.

We have talked to some researchers who work on structure functions, and their codes (again, they are not open to public) typically involve 6 nested for loops (for 3D fields). However, we employ vectorization and 3 loops (instead of 6) in our code; this makes our code approximately 20 times faster than the schemes involving 6 loops. We mention about this in the begining of section titled "Design of the code".
