### 2020-07-10 #2185 @shashwat-bhattacharya
> And changes I suggested in [#2185 (comment)](https://github.com/openjournals/joss-reviews/issues/2185#issuecomment-641971304) are still relevant. At least processors should be changed to cores.

Thanks for the suggestions. We have done the necessary modifications.
