### 2020-07-10 #2185 @shashwat-bhattacharya
> @jedbrown does performance evaluation also include scalability? Test programs included in repository are only 2D so it might be difficult to evaluate scalability claims made in article. I suggested adding 3D test program in [ShubhadeepSadhukhan1993/fastSF#2](https://github.com/ShubhadeepSadhukhan1993/fastSF/issues/2)

Based on the suggestions of @iljah and @jedbrown , we have added 3D test cases (velocity and scalar structure functions). Please refer to `README.md` for details.

Thanks for the suggestions.
