### 2020-07-10 #2185 @shashwat-bhattacharya
@jedbrown @iljah @cpgr We thank you for a comprehensive review of our code. We have implemented your suggestions to the best of our ability. We will welcome further comments, if any. We hope that our work is now suitable for publication in JOSS.
