+++ in/para.yaml
@@ -5,7 +5,7 @@ program:
     scalar_switch: false
   
     #Please select "true" for 2D operations, "false" for 3D operations:
-    2D_switch : true
+    2D_switch : false
 
     #Please select "true" for computing only the longitudinal structure functions, "false" for computing both the transverse and longitudinal structure functions:
     Only_longitudinal: false
@@ -32,7 +32,7 @@ domain_dimension :
 #Please provide the starting order (q1) and the ending order (q2)
 structure_function :
     q1 : 1
-    q2 : 10
+    q2 : 4
 
 #Please enter "true" only if you want to run a test case. WARNING: For test cases, the input fields will be generated by the code. The code will ignore
 # the hdf5 files in the "in" folder. Further,the grid_switch will be automatically set to "true". Thus, the entries against "grid_switch" and "field_procedure" 
