+++ test/test_velocity_2D/in/para.yaml
@@ -1,21 +1,15 @@
 #PARAMETERS FOR COMPUTING THE STRUCTURE FUNCTIONS"
 
 program:
-    #Please enter "true" for computing structure functions as function of lx, ly (for 3D), and lz along with l. If "false is selected, only structure functions as function of l
-    #will be computed.
-    grid_switch: true
-
-    #Please enter "true" for computing scalar structure function, "false" for computing velocity structure function:
+    #Please select "true" for computing scalar structure function, "false" for computing velocity structure function:
     scalar_switch: false
   
-    #Please enter "true" for 2D operations, "false" for 3D operations:
+    #Please select "true" for 2D operations, "false" for 3D operations:
     2D_switch : true
 
-    #Please enter "true" for computing only the longitudinal structure functions, "false" for computing both the transverse and longitudinal structure functions:
+    #Please select "true" for computing only the longitudinal structure functions, "false" for computing both the transverse and longitudinal structure functions:
     Only_longitudinal: false
 
-    #Please enter the number of OpenMP processors
-    Number_of_OpenMP_processors: 1
 
 #Please specify the number of grid points. 
 #Note: Nx - number of points in the x direction, Ny - number of points in the y-direction, Nz - number of points in the z direction.
