### 2020-07-13 ShubhadeepSadhukhan1993/fastSF#1 @cpgr
Changes look great. I'll close this issue and complete my review.
