### 2020-08-01 964b41c
paper: wording edits in summary
