### 2020-08-01 28dd7e2
paper: use "root process" (which has a specific MPI meaning) instead of "master processor"
