### 2020-08-01 #2185 @jedbrown
Thanks for your reviews and my apologies for my greatly delayed reply. I made some copy edits (see PR) and also encountered some further issues that I'd like you to consider.

* It looks like any MPI-1 implementation should be sufficient. I.e., one need not require MPICH.
* Software dependencies should be cited when possible. For example, [this](https://doi.org/10.11578/dc.20180330.1) appears to be the [preferred citation](https://www.hdfgroup.org/2018/06/citations-for-hdf-data-and-software/) for HDF5. In a JOSS paper, it's better to explain the use of the software by name, as in "fastSF uses MPI [cite] for parallelism, HDF5 [cite] via H5SI [cite] for reading gridded field data and writing structure functions, as well as blitz++ [cite] for vectorized computation and yaml-cpp [cite] for reading control parameters."
* It should be possible to use pkg-config or other compile-time configuration to use dependencies from their existing locations. It's poor packaging practice to require dependencies to be in specific directories.
* If there is a memory requirement that each process must be able to redundantly store the entire grid, it should be stated clearly. This becomes an issue if one runs around 1000^3. (Have you considered using MPI shared memory constructs so that processes on a node can share the same read-only data?)
* I think it would be really useful for the application to take its configuration via the command line. This is much more convenient for scripting, in which one might have a loop around calling `fastSF`. You could keep the YAML files for defaults, but I think it's poor practice to not allow the user to configure the task via the command line.
* The input and output files should also be specifiable at run-time.
* Is there a reason why the grid dimensions must be specified up-front instead of being read from the HDF5 files?
* It's usually best practice to write a single HDF5 file with multiple fields instead of one file per field. Would support for such schema be hard to add?
* The supported HDF5 schema should be documented prominently (e.g., in the README). There should be error handling to report what is wrong if a users file does not conform to the schema. If `double` precision is required, this should also be documented.
* It appears that CMake is only required to build `h5si`, but not for `fastSF`. As such, it should not be named as a dependency.
* We usually ask that Python software be `pip` installable. The nearest thing for C++ software in the HPC space would be adding it to [Spack](https://github.com/spack/spack/). This is not required, but does significantly increase likelihood of use and the process tends to fix portability quirks.
* APA style says that journal names should not be abbreviated.
