### 2020-08-02 843d9ea
File existence check is implemented
