### 2020-08-03 284f098
Modified the bib entries and cited dependencies
