+++ paper.bib
@@ -35,7 +35,7 @@
 
 @article{Kolmogorov:Dissipation,
 	Author = {Kolmogorov, Andrey Nikolaevich},
-	Journal = {Dokl Acad Nauk SSSR},
+	Journal = {Doklady Academii Nauk SSSR},
 	Number = {1},
 	Pages = {16--18},
 	Title = {{Dissipation of Energy in Locally Isotropic Turbulence}},
@@ -45,7 +45,7 @@
 
 @article{Kolmogorov:Structure,
 	Author = {Kolmogorov, Andrey Nikolaevich},
-	Journal = {Dokl Acad Nauk SSSR},
+	Journal = {Doklady Academii Nauk SSSR},
 	Number = {4},
 	Pages = {301--305},
 	Title = {{The local structure of turbulence in incompressible viscous fluid for very large Reynolds numbers}},
@@ -55,7 +55,7 @@
 
 @article{She:PRL1993,
 	Author = {She, Zhen-Su and Chen, Shiyi and Doolen, Gary D and Kraichnan, Robert H. and Orszag, Steven A.},
-	Journal = {Phys. Rev. Lett.},
+	Journal = {Physical Review Letters},
 	Month = jan,
 	Pages = {3251--3254},
 	Title = {{Reynolds number dependence of isotropic Navier-Stokes turbulence}},
@@ -65,7 +65,7 @@
 
 @article{Chatterjee:JPDC2018,
 	Author = {Chatterjee, Anando G and Verma, Mahendra K. and Kumar, Abhishek and Samtaney, Ravi and Hadri, Bilel and Khurram, Rooh},
-	Journal = {J. Parallel Distrib. Comput.},
+	Journal = {Journal of Parallel and Distributed Computing},
 	Month = mar,
 	Pages = {77--91},
 	Title = {{Scaling of a Fast Fourier Transform and a pseudo-spectral fluid solver up to 196608 cores}},
@@ -75,7 +75,7 @@
 
 @article{Yeung:PF2005,
 	Author = {Yeung, P. K. and Donzis, Diego A. and Sreenivasan, Katepalli R.},
-	Journal = {Phys. Fluids},
+	Journal = {Physics of Fluids},
 	Month = aug,
 	Number = {8},
 	Pages = {081703},
@@ -86,7 +86,7 @@
 
 @article{Kaneda:PF2003,
 	Author = {Ishihara, Takashi and Yokokawa, Mitsuo and Itakura, Ken'ichi and Uno, Atsuya},
-	Journal = {Phys. Fluids},
+	Journal = {Physics of Fluids},
 	Month = feb,
 	Number = {2},
 	Pages = {L21},
@@ -99,7 +99,7 @@
 	Author = {Gotoh, Toshiyuki and Fukayama, Daigen and Nakano, Tohru},
 	Date-Added = {2019-09-25 13:20:40 +0530},
 	Date-Modified = {2019-09-25 13:20:40 +0530},
-	Journal = {Phys. Fluids},
+	Journal = {Phys of Fluids},
 	Month = mar,
 	Number = {3},
 	Pages = {1065--1081},
@@ -110,7 +110,7 @@
 
 @article{Ray:NJP2008,
 	Author = {Ray, Samriddhi Sankar and Mitra, Dhrubaditya and Pandit, R.},
-	Journal = {New J. Phys.},
+	Journal = {New Journal of Physics},
 	Month = mar,
 	Number = {3},
 	Pages = {033003},
@@ -121,7 +121,7 @@
 
 @article{Biferale:NJP2004,
 	Author = {Biferale, Luca and Cencini, M. and Lanotte, A. S. and Sbragaglia, M. and Toschi, Federico},
[…]
