+++ paper.bib
@@ -222,6 +222,7 @@
 
 @misc{H5SI_Web,
   title = {H5SI},
+  auther={Chatterjee, Anando G and Chatterjee, Soumyadeep}
   howpublished = {\url{https://github.com/anandogc/h5si}},
   note = {Accessed: 2020-08-03}
 }
