+++ paper.bib
@@ -222,9 +222,10 @@
 
 @misc{H5SI_Web,
   title = {H5SI},
-  auther={Chatterjee, Anando G and Chatterjee, Soumyadeep}
+  auther={Chatterjee, Anando G and Chatterjee, Soumyadeep},
   howpublished = {\url{https://github.com/anandogc/h5si}},
-  note = {Accessed: 2020-08-03}
+  note = {Accessed: 2020-08-03},
+  year = {2020}
 }
 
 @misc{MPICH_Web,
