+++ paper.bib
@@ -220,12 +220,13 @@
   doi = {10.1007/3-540-49372-7_24}
 }
 
-@misc{H5SI_Web,
-  title = {H5SI},
-  auther={Chatterjee, Anando G and Chatterjee, Soumyadeep},
-  howpublished = {\url{https://github.com/anandogc/h5si}},
-  note = {Accessed: 2020-08-03},
-  year = {2020}
+@misc{H5SI_web,
+	title = {H5SI},
+	author = {Chatterjee, Anando and Chatterjee, Soumyadeep},
+	url = {https://https://github.com/anandogc/h5si},
+	howpublished = {[Computer Software] \url{https://github.com/anandogc/h5si}},
+	year = {2020},
+	month = {feb}
 }
 
 @misc{MPICH_Web,
