### 2020-08-03 8abb54d
Update README.md
Removed CMake from the list of dependencies.
