### 2020-08-17 f3303b9
command line input implemented
