### 2020-08-17 aa60878
Error message and checklist for input file is added
