### 2020-08-18 0e07640
Input field compatibility check is implemented
