### 2020-08-18 a974241
Updated docs, implemented px in para.yaml
