+++ test/test_velocity_2D/in/para.yaml
@@ -10,6 +10,9 @@ program:
     #Please select "true" for computing only the longitudinal structure functions, "false" for computing both the transverse and longitudinal structure functions:
     Only_longitudinal: false
 
+    #Please enter the number of processors in x direction:
+    Processors_X: 1
+
 
 #Please specify the number of grid points. 
 #Note: Nx - number of points in the x direction, Ny - number of points in the y-direction, Nz - number of points in the z direction.
