### 2020-08-18 c1aa7d1
Updated Makefile and the source code
