+++ src/fastSF.cc
@@ -770,8 +770,8 @@ void write_SFs() {
 *************************************************************************************************************************************
 */
 void test_cases() {
-    cout<<"\nCOMMENCING TESTING OF THE CODE.\n";
     if(rank_mpi==0){
+        cout<<"\nCOMMENCING TESTING OF THE CODE.\n";
         if (scalar_switch){
             if (two_dimension_switch){
                 SCALAR_TEST_CASE_2D();
