+++ in/para.yaml
@@ -38,7 +38,6 @@ structure_function :
     q2 : 4
 
 #Please enter "true" only if you want to run a test case. WARNING: For test cases, the input fields will be generated by the code. The code will ignore
-# the hdf5 files in the "in" folder. Further,the grid_switch will be automatically set to "true". Thus, the entries against "grid_switch" and "field_procedure" 
-# will be overriden. It is strongly recommended not to use a grid not finer than 32^3.
+# the hdf5 files in the "in" folder. Further,the grid_switch will be automatically set to "true". It is strongly recommended not to use a grid finer than 32^3.
 test :
     test_switch : true
