### 2020-08-18 bc76013
Update README.md
Added memory requirements.
