### 2020-08-18 76bdf00
Input and output files specifiable during runtime
