### 2020-08-18 09fab05
filename from command line bug fixed and checklist message from one process is done.
