+++ src/fastSF.cc
@@ -364,6 +364,60 @@ int P;
  */
 int px;
 
+/**
+ ********************************************************************************************************************************************
+ * \brief   This variable stores the name for the input file for velocity field along x axis.
+ *
+ ********************************************************************************************************************************************
+ */
+string UName="U.V1r";
+
+/**
+ ********************************************************************************************************************************************
+ * \brief   This variable stores the name for the input file for velocity field along y axis.
+ *
+ ********************************************************************************************************************************************
+ */
+string VName="U.V2r";
+/**
+ ********************************************************************************************************************************************
+ * \brief   This variable stores the name for the input file for velocity field along z axis.
+ *
+ ********************************************************************************************************************************************
+ */
+string WName="U.V3r";
+
+/**
+ ********************************************************************************************************************************************
+ * \brief   This variable stores the name for the input file for scalar field.
+ *
+ ********************************************************************************************************************************************
+ */
+ string TName="T.Fr";
+
+ /**
+ ********************************************************************************************************************************************
+ * \brief   This variable stores the name for the output file for parallel structure function of velocity field.
+ *
+ ********************************************************************************************************************************************
+ */
+string SF_Grid_pll_name = "SF_Grid_pll";
+
+/**
+ ********************************************************************************************************************************************
+ * \brief   This variable stores the name for the output file for perpendicular structure function of velocity field.
+ *
+ ********************************************************************************************************************************************
+ */
+string SF_Grid_perp_name = "SF_Grid_perp";
+
+/**
+ ********************************************************************************************************************************************
+ * \brief   This variable stores the name for the output file for structure function of scalar field.
+ *
+ ********************************************************************************************************************************************
+ */
+ string SF_Grid_scalar_name = "SF_Grid_scalar";
 
 
 /**
@@ -392,9 +446,7 @@ int main(int argc, char *argv[]) {
     get_Inputs();
     //Overwrite from command line
     int option;
-    string UName="U.V1r", VName="U.V2r", WName="U.V3r", TName="T.Fr";
[…]
