+++ src/fastSF.cc
@@ -58,9 +58,9 @@ using namespace blitz;
 
 //Function declarations
 void get_Inputs(); 
-void write_3D(Array<double,3>, string, int);
-void write_4D(Array<double,4>, string, int);
-void read_2D(Array<double,2>, string, string);
+void write_3D(Array<double,3>, string);
+void write_4D(Array<double,4>, string);
+void read_2D(Array<double,2>, string, string, string);
 string int_to_str(int);
 bool str_to_bool(string);
 void VECTOR_TEST_CASE_3D();
@@ -69,7 +69,7 @@ void SCALAR_TEST_CASE_2D();
 void SCALAR_TEST_CASE_3D();
 void compute_time_elapsed(timeval, timeval, double&);
 
-void read_3D(Array<double,3>, string, string);
+void read_3D(Array<double,3>, string, string, string);
 
 
 void SFunc2D(Array<double,2>, Array<double,2>);
@@ -704,21 +704,21 @@ void Read_fields(string UName, string VName, string WName, string TName) {
         }
         if (two_dimension_switch){
             if (scalar_switch) {
-                read_2D(T_2D,"in/", TName);
+                read_2D(T_2D,"in/", TName, TName);
             }
             else {
-                read_2D(V1_2D,"in/", UName);
-                read_2D(V3_2D,"in/", WName);
+                read_2D(V1_2D,"in/", UName, UName);
+                read_2D(V3_2D,"in/", WName, WName);
             }
         }
         else{
             if (scalar_switch) {
-                read_3D(T, "in/", TName);
+                read_3D(T, "in/", TName, TName);
             }
             else {
-                read_3D(V1, "in/", UName);
-                read_3D(V2, "in/", VName);
-                read_3D(V3, "in/", WName);
+                read_3D(V1, "in/", UName, UName);
+                read_3D(V2, "in/", VName, VName);
+                read_3D(V3, "in/", WName, WName);
             }
         }
     }
@@ -808,36 +808,33 @@ void calc_SFs() {
 void write_SFs(string SF_Grid_pll_name, string SF_Grid_perp_name, string SF_Grid_scalar_name) {
     if (rank_mpi==0){
         mkdir("out",0777);
-        int p1 = q1;
-        while (p1 <= q2) {
-            string name = int_to_str(p1);
-            if (two_dimension_switch) {
-                cout<<"\nWriting "<<p1<<" order SF as function of lx and lz\n";
-                if (scalar_switch){
-                    write_3D(SF_Grid2D_scalar, SF_Grid_scalar_name+name, p1);
-                }
-                else {
-                    write_3D(SF_Grid2D_pll, SF_Grid_pll_name+name, p1);    
-                    if (not longitudinal) {
-                        write_3D(SF_Grid2D_perp, SF_Grid_perp_name+name, p1);
-                    }
-                }
-                cout<<"\nWriting completed\n";
+       
+
+        if (two_dimension_switch) {
+            
+            if (scalar_switch){
+                write_3D(SF_Grid2D_scalar, SF_Grid_scalar_name);
             }
             else {
-                cout<<"\nWriting "<<p1<<" order SF as function of lx, ly, and ly\n";
-                if (scalar_switch){
-                    write_4D(SF_Grid_scalar, SF_Grid_scalar_name+name, p1);
+                write_3D(SF_Grid2D_pll, SF_Grid_pll_name);    
+                if (not longitudinal) {
+                    write_3D(SF_Grid2D_perp, SF_Grid_perp_name);
                 }
[…]
