+++ in/para.yaml
@@ -40,4 +40,4 @@ structure_function :
 #Please enter "true" only if you want to run a test case. WARNING: For test cases, the input fields will be generated by the code. The code will ignore
 # the hdf5 files in the "in" folder. Further,the grid_switch will be automatically set to "true". It is strongly recommended not to use a grid finer than 32^3.
 test :
-    test_switch : false
+    test_switch : true
