### 2020-08-25 9821a64
Command line parsing moved to the read_input function; documentation updated
