+++ src/fastSF.cc
@@ -57,7 +57,7 @@ using namespace std;
 using namespace blitz;
 
 //Function declarations
-void get_Inputs(); 
+void get_Inputs(int argc, char* argv[]); 
 void write_3D(Array<double,3>, string);
 void write_4D(Array<double,4>, string);
 void read_2D(Array<double,2>, string, string, string);
@@ -95,10 +95,10 @@ void SF_scalar_3D(Array<double,3>);
 
 void SF_scalar_2D(Array<double,2>);
 
-void Read_fields(string, string, string, string);
+void Read_fields();
 void resize_SFs();
 void calc_SFs();
-void write_SFs(string, string, string);
+void write_SFs();
 void test_cases();
 void show_checklist();
 
@@ -366,7 +366,7 @@ int px;
 
 /**
  ********************************************************************************************************************************************
- * \brief   This variable stores the name for the input file for velocity field along x axis.
+ * \brief   This variable stores the name for the input file for the x-component of the velocity field.
  *
  ********************************************************************************************************************************************
  */
@@ -374,14 +374,14 @@ string UName="U.V1r";
 
 /**
  ********************************************************************************************************************************************
- * \brief   This variable stores the name for the input file for velocity field along y axis.
+ * \brief   This variable stores the name for the input file for the y-component of the velocity field.
  *
  ********************************************************************************************************************************************
  */
 string VName="U.V2r";
 /**
  ********************************************************************************************************************************************
- * \brief   This variable stores the name for the input file for velocity field along z axis.
+ * \brief   This variable stores the name for the input file for the z-component of the velocity field.
  *
  ********************************************************************************************************************************************
  */
@@ -423,7 +423,10 @@ string SF_Grid_perp_name = "SF_Grid_perp";
 /**
  ********************************************************************************************************************************************
  * \brief   The main function of the "fastSF".
- 
+ *
+ * \param   argc is the number of command-line arguments passed by the user including the name of the program.
+ * \param   argv is the array of character pointers listing all the arguments.
+ *
  ********************************************************************************************************************************************
  */
 int main(int argc, char *argv[]) {
@@ -442,92 +445,11 @@ int main(int argc, char *argv[]) {
     double elapsedt=0.0;
     double elapsepdt=0.0;
     
+    //Get the input parameters
+    get_Inputs(argc, argv);
     
-    get_Inputs();
-    //Overwrite from command line
-    int option;
     
-    while ((option=getopt(argc, argv, "X:Y:Z:1:2:x:y:z:l:d:p:t:s:U:V:W:S:P:L:M:"))!=-1){
-    	switch(option){
-    		case 'X':
-    			Nx=std::stoi(optarg);
-    			break;
-    		case 'Y':
-    			Ny=std::stoi(optarg);
-    			break;
-    		case 'Z':
-    			Nz=std::stoi(optarg);
-    			break;
-    		case 'x':
-    			Lx=std::stod(optarg);
-    			break;
-    		case 'y':
-    			Ly=std::stod(optarg);
-    			break;
-    		case 'z':
-    			Lz=std::stod(optarg);
-    			break;
-    		case 'p':
-    			px=std::stod(optarg);
-    			break;
-    		case '1':
-    			q1=std::stod(optarg);
-    			break;
-    		case '2':
[…]
