+++ README.md
@@ -114,10 +114,48 @@ Finally, for visualization purpose, the python script `test/test.py` is invoked.
 
 This section provides a detailed procedure to execute `fastSF` for a given velocity or scalar field.
 
+### i) Files Required and HDF5 Schema:
+
+All the files storing the input fields should be inside the `in` folder. All the input fields and datasets are required to be precisely as per the following schema.
+
+#### For two dimensional fields
+
+For vector field, two `hdf5` files are required:
+
+1. The hdf5 file storing the x-component of the vector / velocity field. Unless specified by the user via command line arguments, the hdf5 file should be named as `U.V1r.h5`, with a 2D dataset `U.V1r` of dimensions (`Nx,Nz`) storing real double-precision floating point values. 
+2. The hdf5 file storing the z-component of the vector / velocity field. Unless specified by the user via command line arguments, the hdf5 file should be named as `U.V3r.h5`, with a 2D dataset `U.V3r` of dimensions (`Nx,Nz`) storing real double-precision floating point values. 
+
+For scalar field, For vector field, one `hdf5` file is required:
+
+1. The hdf5 file storing the scalar field. Unless specified by the user via command line arguments, the hdf5 file should be named as `T.Fr.h5`, with a 2D dataset `T.Fr` of dimensions (`Nx,Nz`) storing real double-precision floating point values.
+
+
+#### For three dimensional fields
+
+For vector field, three `hdf5` files are required:
+
+1. The hdf5 file storing the x-component of the vector / velocity field. Unless specified by the user via command line arguments, the hdf5 file should be named as `U.V1r.h5`, with a 3D dataset `U.V1r` of dimensions (`Nx,Ny,Nz`) storing real double-precision floating point values. 
+2. The hdf5 file storing the y-component of the vector / velocity field. Unless specified by the user via command line arguments, the hdf5 file should be named as `U.V2r.h5`, with a 3D dataset `U.V2r` of dimensions (`Nx,Ny,Nz`) storing real double-precision floating point values.
+3. The hdf5 file storing the z-component of the vector / velocity field. Unless specified by the user via command line arguments, the hdf5 file should be named as `U.V3r.h5`, with a 3D dataset `U.V3r` of dimensions (`Nx,Ny,Nz`) storing real double-precision floating point values.
+
+For scalar field, For vector field, one `hdf5` file is required:
+
+1. The hdf5 file storing the scalar field. Unless specified by the user via command line arguments, the hdf5 file should be named as `T.Fr.h5`, with a 3D dataset `T.Fr` of dimensions (`Nx,Ny,Nz`) storing real double-precision floating point values.
+
+**IMPORTANT:** Please note the following:
+
+* The input files can have different names than the ones specified above, but it needs to be ensured that the hdf5 file name and the dataset name are the same (except that the dataset name does not have `.h5` extension. The input file names need to be specified via command-line arguments (explained later).
+
+* The user must ensure that the size of the input field datasets match with the size specified in the `para.yaml`, otherwise the code will likely throw a segmentation error. Even if the code does not throw a segmentation error, the calculations will be erroneous. 
+
+### ii) `para.yaml` details
+
+The user can specify the relevant parameters via command line or via parameters file. If no command-line options are given, the entries in the parameters file will be taken. First, we will explain how to use the parameters file, and then we will proceed to command-line options.
+
[…]
