+++ paper.md
@@ -40,7 +40,7 @@ Turbulence is a complex phenomenon in fluid dynamics involving nonlinear interac
 We are not aware of any other open soure or commercial packages for computing structure functions; prior studies have relied on in-house software that was never publicly released.  As an open source package, `fastSF` provides a standard high-performance implementation and thus facilitates wider use of structure functions.
 
 ``fastSF`` uses MPI [@Pacheco:book:PP] for parallelism, HDF5 [@HDF5_web] via H5SI [@H5SI_web] for reading gridded field data and writing structure functions, as well as blitz++ [@Blitz_web] for vectorized computation and yaml-cpp [@YAML_web] for reading control parameters.
- In the next section, we will briefly discuss the performance and scaling of ``fastSF`` in a Cray XC40 system.
+ In the next section, we will briefly explain the velocity and scalar structure functions in turbulent flows.
 
 
 
