### 2020-08-30 #2185 @shashwat-bhattacharya
@jedbrown I would like to thank you again for your useful recommendations. We have implemented them to the best of our ability as follows:

* We mention in the README and in the revised paper that any MPI-1 implementation will suffice.
* We have cited all the software dependencies in the revised paper.
* In the updated code, it is possible to use pkg-config to use yaml-cpp and blitz++ libraries. However, we could not find pkg-config files for hdf5 and h5si libraries.
* We have added a section on estimating the memory requirement in the README file. We agree that making all the processors store the entire grid makes the code very memory intensive. However, due to the nature of our algorithm, storing the grid in one master processor will require complex communications between the processers which can slow down the code. Moreover, computing the structure functions on very large grids (above 512^3) is not really required because structure functions are generally important in the intermediate scales. Thus, one can coarsen the input data to 512^3 grid or lower resolutions and still get accurate results for the structure functions with less memory requirement.
* We have implemented the facility of accepting inputs via command line. If the user does not specify a particular parameter via command line, then that parameter value is taken from para.yaml file. It should now be more convenient for the user for scripting.
* The input and the output names are now specifiable at run-time. If these names are not passed, then the default naming convention (as it was before) will be followed.
* The grid dimensions are read from the parameters file because we first resize the field array before feeding the input data to it from the hdf5 file. Not doing so can cause segmentation faults. This seems to be a minor drawback of either h5si or Blitz++ library.
* The revised code stores the output in two hdf5 files for velocity structure functions (one for longitudinal and the other for transverse) and one hdf5 file for scalar structure functions. Each hdf5 file has multiple datasets corresponding to different orders of the structure function. Thus, each dataset stores the structure functions of a particular order. Please note that in the earlier code, different orders of structure functions were stored in separate hdf5 files. Thus, the output file schema is more compact in the revised code.
* The revised README file has clear instructions on the supported HDF5 schema. Please note that currently the code accepts only double-precision values. If there is an error in the user's schema, the code will display a checklist for the schema.
* CMake has been removed from the list of dependencies.
* The idea of Spack looks good, and we are planning to add `fastSF` to Spack for later versions. We thank you for this suggestion.
* The journal names are written in full in the bibliography of the revised paper.

We thank you again for these useful suggestions. They have definitely improved the code. I hope the paper is now suitable for publication. We will welcome further comments, if any.
