### 2020-11-09 #2185 @shashwat-bhattacharya
Hi @jedbrown , we are awaiting your response.
We are using this code for our subsequent work and thus it will be helpful if this paper is published soon. Please let us know if anything further needs to be done to the code or the manuscript such that it is suitable for publication in JOSS.
