### 2020-11-09 #2185 @jedbrown
Thanks for your updates and sorry about my delinquency getting to this.

* Package managers often install pkg-config for HDF5 (e.g., Debian's [libhdf5-dev](https://packages.debian.org/buster/amd64/libhdf5-dev/filelist)). The repository has [a stub](https://bitbucket.hdfgroup.org/projects/HDFFV/repos/hdf5/browse/config/cmake/libhdf5.pc.in) but you're correct that it's often not installed. In the old days, we all used `./configure --with-hdf5-dir=/path/to/hdf5` or if using a raw Makefile, `make HDF5_DIR=/path/to/hdf5` and have the makefile calculate appropriate compile and link flags. I understand you have a single-file project, but asking users to edit your makefile is clumsy and avoidable if you just use variables.

* HDF5-1.8 has reach end-of-life. The last release was 1.8.21 in June 2018 "HDF5 1.8 will not be supported after the May 2020 release". I think new software should use supported versions of dependencies unless there is a really strong reason not to (e.g., license incompatibility).

* HDF5 has solid support for reading metadata (including shapes) before reading the array data. It seems like h5si has a `shape()` method that may do what you need: https://github.com/anandogc/h5si/blob/master/trunk/lib/h5dataset.h#L104 (There don't seem to be any docs or nontrivial tests so who knows if it works.)

* It's good practice for command-line executables to respond to `-h` and/or `--help` to document their options.

* Why do the input files need to have special names (according to the README)? Why not just scan one or more input files looking for datasets with the specified names? It just seems like an odd redundancy that causes friction compared to how everyone else creates HDF5 files.

> **IMPORTANT:** Please note the following:
>
> * The input files can have different names than the ones specified above, but it needs to be ensured that the hdf5 file name and the dataset name are the same (except that the dataset name does not have `.h5` extension. The input file names need to be specified via command-line arguments (explained later).
