### 2020-11-09 #2185 @jedbrown
* The `\mathbf l` in your paper looks very much like`1` in our font. Could you use `\bm l` (or use `\ell`) to avoid this ambiguity?
* "Phys of Fluids" -> "Physics of Fluids"
* Capitalization in the yaml-cpp reference
* I don't think the discussion of six vs three nested loops is a great way to describe what's going on. I'd just say you use vectorization. I also don't think this implementation as described is anywhere close to optimal. You're currently reading all the data for each value of `l`, but you could batch together a tile of those `l` so that you make many fewer passes through the global data. This has similar structure to GEMM and convolutions, both of which can achieve near-peak floating point. I predict that if you estimate a flops rate for your code (or use a profiler with hardware perf counters) you'll find that it's bottlenecked on memory bandwidth despite the surplus of flops. It's okay if you don't address this, but in that case, you can just say we use this vectorized computation strategy without making specific claims of high performance.

Apart from that, it looks good modulo the software distribution/best-practices points above, after which we'll be ready to archive.
