### 2020-11-17 59bed98
input reading from file is updated. Now Nx, Ny, Nz will come from input data while reading from data files
