+++ src/fastSF.cc
@@ -82,7 +82,6 @@ void SFunc3D(Array<double,3>, Array<double,3>, Array<double,3>);
 
 
 void SFunc_long_3D(Array<double,3>, Array<double,3>, Array<double,3>);
-bool compatibility_check(h5::Dataset, int, int, int);
 void Read_Init(Array<double,2>&, Array<double,2>&);
 void Read_Init(Array<double,3>&, Array<double,3>&, Array<double,3>&);
 void Read_Init(Array<double,2>&);
@@ -102,7 +101,10 @@ void write_SFs();
 void test_cases();
 void show_checklist();
 
-
+bool compare(Array<int,1> , Array<int,1> );
+void calculate_grid_spacing();
+void resize_input();
+void get_input_shape(string , string , string , Array<int,1>&);
 
 /**
  ********************************************************************************************************************************************
@@ -449,18 +451,7 @@ int main(int argc, char *argv[]) {
     get_Inputs(argc, argv);
     
     
-    //Specify the values of dx, dy, and dz
-    if (Nx==1){dx=0;}
-    else{
-      dx=Lx/double(Nx-1);}
-    if (Ny==1){dy=0;}
-    else{
-      dy=Ly/double(Ny-1);}
-    if (Nz==1){dz=0;}
-    else{
-      dz=Lz/double(Nz-1);
-  	}
-
+    
   	if (rank_mpi==0) {
     	cout<<"\nNumber of processors in x direction: "<<px<<endl;
     	if (two_dimension_switch) {
@@ -505,8 +496,14 @@ int main(int argc, char *argv[]) {
         MPI_Finalize();
         exit(1);
     } 
+
+
+
     //Resizing the input fields
     Read_fields();
+
+
+
 
     //Resize the structure function array according to the type of inputs
     resize_SFs();
@@ -549,6 +546,27 @@ int main(int argc, char *argv[]) {
 }
 
 /**
+******************************************************************************************************************************
+*\brief  Function to calculate dx, dy and dz
+*
+******************************************************************************************************************************
+*/
+void calculate_grid_spacing(){
+	//Specify the values of dx, dy, and dz
+    if (Nx==1){dx=0;}
+    else{
+      dx=Lx/double(Nx-1);}
+    if (Ny==1){dy=0;}
+    else{
+      dy=Ly/double(Ny-1);}
+    if (Nz==1){dz=0;}
+    else{
+      dz=Lz/double(Nz-1);
+  	}
+}
+
+
+/**
 *************************************************************************************************************************************
 *\brief     Function to covert string to bool
 *           
@@ -568,21 +586,72 @@ bool str_to_bool(string s){
 	}
 }
 
+
 /**
-*************************************************************************************************************************************
-*\brief     Function to generate or read the input fields.
+********************************************************************************************************************************
+*\brief    Function to check the file does exist or not. Furthermore it saves the shape of the dataset.
 *
+*\param fold is data path
+*\param file is the file name
+*\param dset is the dataset name
+*\param s is array in which shape will be saved
 *
-*\param     UName is the name of the hdf5 file and dataset storing the x-component of velocity / vector field. 
-*\param     VName is the name of the hdf5 file and dataset storing the y-component of velocity / vector field.
-*\param     WName is the name of the hdf5 file and dataset storing the z-component of velocity / vector field.
-*\param     TName is the name of the hdf5 file and dataset storing the scalar field.
-* 
-*************************************************************************************************************************************
[…]
