### 2020-11-18 d4078bc
Update fastSF.cc
