### 2020-11-18 84cecaa
Minor changes in documentation
