+++ src/fastSF.cc
@@ -398,6 +398,41 @@ string WName="U.V3r";
  */
  string TName="T.Fr";
 
+/**
+ ********************************************************************************************************************************************
+ * \brief   This variable stores the name for the dataset for the x-component of the velocity field.
+ *
+ ********************************************************************************************************************************************
+ */
+string UdName="U.V1r";
+
+
+/**
+ ********************************************************************************************************************************************
+ * \brief   This variable stores the name for the dataset for the y-component of the velocity field.
+ *
+ ********************************************************************************************************************************************
+ */
+string VdName="U.V2r";
+
+
+/**
+ ********************************************************************************************************************************************
+ * \brief   This variable stores the name for the dataset for the z-component of the velocity field.
+ *
+ ********************************************************************************************************************************************
+ */
+string WdName="U.V3r";
+
+
+/**
+ ********************************************************************************************************************************************
+ * \brief   This variable stores the name for the dataset of the scalar field.
+ *
+ ********************************************************************************************************************************************
+ */
+string TdName="T.Fr";
+
  /**
  ********************************************************************************************************************************************
  * \brief   This variable stores the name for the output file for parallel structure function of velocity field.
@@ -720,14 +755,14 @@ void Read_fields() {
         
         if (two_dimension_switch){
             if (scalar_switch) {
-            	get_input_shape("in/", TName, TName, s1);
+            	get_input_shape("in/", TName, TdName, s1);
                 resize_input();
                 calculate_grid_spacing();
-                read_2D(T_2D,"in/", TName, TName);
+                read_2D(T_2D,"in/", TName, TdName);
             }
             else {
-            	get_input_shape("in/", UName, UName, s1);
-                get_input_shape("in/", WName, WName, s2);
+            	get_input_shape("in/", UName, UdName, s1);
+                get_input_shape("in/", WName, WdName, s2);
                 
                 if (!compare(s1,s2)){
                 	if (rank_mpi==0){
@@ -741,23 +776,23 @@ void Read_fields() {
                 
                 resize_input();
                 calculate_grid_spacing();
-                read_2D(V1_2D,"in/", UName, UName);
-                read_2D(V3_2D,"in/", WName, WName);
+                read_2D(V1_2D,"in/", UName, UdName);
+                read_2D(V3_2D,"in/", WName, WdName);
             }
         }
         else{
         	
             if (scalar_switch) {
-            	get_input_shape("in/", TName, TName, s1);
+            	get_input_shape("in/", TName, TdName, s1);
             	resize_input();
             	calculate_grid_spacing();
-                read_3D(T, "in/", TName, TName);
+                read_3D(T, "in/", TName, TdName);
             }
             else {
             	
-            	get_input_shape("in/", UName, UName, s1);
-            	get_input_shape("in/", VName, VName, s2);
-            	get_input_shape("in/", WName, WName, s3);
+            	get_input_shape("in/", UName, UdName, s1);
+            	get_input_shape("in/", VName, VdName, s2);
[…]
