+++ paper.md
@@ -46,38 +46,38 @@ We are not aware of any other open soure or commercial packages for computing st
 
 # Velocity and scalar structure functions
 
-We denote the velocity and scalar fields using $\mathbf{u}$ and $\theta$  respectively. The velocity difference between any two points $\mathbf{r}$ and $\mathbf{r}+\boldsymbol{l}$ is $\delta \mathbf{u} = \mathbf{u(r}+ \boldsymbol{l)}-\mathbf{u(r)}$. The difference in the parallel components of the velocity field along $\boldsymbol{l}$ is $\delta u_\parallel=\delta \mathbf{u}\cdot \hat{\boldsymbol{l}}$.  The corresponding difference in the perpendicular component is $\delta u_\perp= |\delta \mathbf{u} - \delta u_\parallel \hat{\mathbf{l}}|$. Assuming statistical homogeneity, we define the longitudinal velocity structure functions of order $q$ as
-$$ S_q^{u_\parallel}(\mathbf{l}) = \langle (\delta u_\parallel)^q \rangle = \langle [\{\mathbf{u(r+l)}-\mathbf{u(r)}\}\cdot \hat{\mathbf{l}}]^q \rangle, \quad \quad (1)$$ 
+We denote the velocity and scalar fields using $\boldsymbol{u}$ and $\theta$  respectively. The velocity difference between any two points $\boldsymbol{r}$ and $\boldsymbol{r}+\boldsymbol{l}$ is $\delta \boldsymbol{u} = \boldsymbol{u(r}+ \boldsymbol{l)}-\boldsymbol{u(r)}$. The difference in the parallel components of the velocity field along $\boldsymbol{l}$ is $\delta u_\parallel=\delta \boldsymbol{u}\cdot \hat{\boldsymbol{l}}$.  The corresponding difference in the perpendicular component is $\delta u_\perp= |\delta \boldsymbol{u} - \delta u_\parallel \hat{\boldsymbol{l}}|$. Assuming statistical homogeneity, we define the longitudinal velocity structure functions of order $q$ as
+$$ S_q^{u_\parallel}(\boldsymbol{l}) = \langle (\delta u_\parallel)^q \rangle = \langle [\{\boldsymbol{u(r+l)}-\boldsymbol{u(r)}\}\cdot \hat{\boldsymbol{l}}]^q \rangle, \quad \quad (1)$$ 
 and the transverse velocity structure functions of order 
 $q$ as 
-$$ S_q^{u_\perp}(\mathbf{l}) = \langle (\delta u_\perp)^q \rangle = \langle |\delta \mathbf{u} - \delta u_\parallel \hat{\mathbf{l}}|^q \rangle. \quad \quad (2)$$ 
+$$ S_q^{u_\perp}(\boldsymbol{l}) = \langle (\delta u_\perp)^q \rangle = \langle |\delta \boldsymbol{u} - \delta u_\parallel \hat{\boldsymbol{l}}|^q \rangle. \quad \quad (2)$$ 
 Here, $\langle \cdot \rangle$ denotes ensemble averaging. Similarly, we can define the scalar structure functions for the scalar field as 
-$$ S_q^\theta(\mathbf{l}) = \langle (\delta \theta)^q\rangle = \langle [\theta (\mathbf{r+l}) - \theta(\mathbf{r})]^q \rangle. \quad \quad(3)$$
+$$ S_q^\theta(\boldsymbol{l}) = \langle (\delta \theta)^q\rangle = \langle [\theta (\boldsymbol{r+l}) - \theta(\boldsymbol{r})]^q \rangle. \quad \quad(3)$$
 
[…]
