+++ paper.bib
@@ -99,7 +99,7 @@
 	Author = {Gotoh, Toshiyuki and Fukayama, Daigen and Nakano, Tohru},
 	Date-Added = {2019-09-25 13:20:40 +0530},
 	Date-Modified = {2019-09-25 13:20:40 +0530},
-	Journal = {Phys of Fluids},
+	Journal = {Physics of Fluids},
 	Month = mar,
 	Number = {3},
 	Pages = {1065--1081},
