+++ paper.md
@@ -59,7 +59,7 @@ For isotropic turbulence (in addition to being homogeneous), the structure funct
 In the next section, we provide a brief description of the code.
 
 # Design of the Code
-First we present a sketch of the structure function computation for the velocity structure functions.  Typical structure function computations [Eqs (1-3)] in literature involve calculation of the velocity difference using loops over $\boldsymbol{r}$ and $\boldsymbol{l}$. These computations require six nested `for` loops for 3D fields that makes the computations very expensive for large grids. In our code, we employ vectorization and loops over only $\boldsymbol{l}$, thus requiring three loops instead of six for 3D fields. The new algorithm enhances the performance approximately 20 times over the earlier schemes due to vectorization.  In the following, we provide the algorithm for structure function computation for a 2D velocity field.
+First we present a sketch of the structure function computation for the velocity structure functions.  We employ vectorization and loops over $\boldsymbol{l}$, thus requiring three loops 3D fields. In the following, we provide the algorithm for structure function computation for a 2D velocity field.
 
 **Pseudo-code**
 
@@ -93,7 +93,9 @@ First we present a sketch of the structure function computation for the velocity
 
 ![The velocity difference $\delta \boldsymbol{u}(\boldsymbol{l})$ is computed by taking the difference between two points with the same indices in the pink and the green subdomains. For example, $\boldsymbol{u}(\boldsymbol{l}) - \boldsymbol{u}(0,0) = \boldsymbol{u}_B - \boldsymbol{u}_A$, where $B$ and $A$ are the origins of the green and the pink subdomains. This feature enables vecotrization of the computation. \label{Schematic}](docs/figs/Schematic.png)
 
-Since $S_q^u(\boldsymbol{l})$ is important for intermediate scales (inertial range) only, we vary $\boldsymbol{l}$ upto half the domain size, that is, upto ($L_x/2, L_z/2$), to save computational cost. The $\boldsymbol{l}$'s are divided among MPI processors along $x$ and $z$ directions. Each MPI processor computes the structure functions for the points assigned to it and has access to the entire input data. Thus, we save considerable time that would otherwise be spent on communication between the processors during the calculation of the velocity difference. After computing the structure function for a given $\boldsymbol{l}$, each processor communicates the result to the root process, which stores the $S_q^{u_\parallel}(\boldsymbol{l})$ and $S_q^{u_\perp}(\boldsymbol{l})$ arrays.
+Since $S_q^u(\boldsymbol{l})$ is important for intermediate scales (inertial range) only, we vary $\boldsymbol{l}$ upto half the domain size, that is, upto ($L_x/2, L_z/2$), to save computational cost. The $\boldsymbol{l}$'s are divided among MPI processors along $x$ and $z$ directions. Each MPI processor computes the structure functions for the points assigned to it and has access to the entire input data. 
+%Thus, we save considerable time that would otherwise be spent on communication between the processors during the calculation of the velocity difference. 
+After computing the structure function for a given $\boldsymbol{l}$, each processor communicates the result to the root process, which stores the $S_q^{u_\parallel}(\boldsymbol{l})$ and $S_q^{u_\perp}(\boldsymbol{l})$ arrays.
 
 It is clear from Fig. \ref{Schematic} that the sizes of the pink or green subdomains are $(L_x-l_x)(L_z-l_z)$, which are function of $\boldsymbol{l}$'s.  This function decreases with increasing $\boldsymbol{l}$ leading to larger computational costs for small $l$ and less cost of larger $l$.   Hence, a straightforward division of the domain among the processors along $x$ and $z$ directions will lead to a load imbalance.   Therefore, we assign both large and small $\boldsymbol{l}$'s to each processor to achieve equal load distribution. We illustrate the above idea  using the following example.
 
