+++ paper.bib
@@ -236,7 +236,7 @@
 }
 
 @article{YAML_web,
-  title={yaml-cpp,” A YAML parser and emitter for C++,“},
+  title={{yaml-cpp,” A YAML parser and emitter for C++,“}},
   author={Beder, Jesse},
   journal={URL: https://github. com/jbeder/yaml-cpp},
   year={2019}
