### 2020-12-07 41f59dc
No need to specify Nx, Ny, Nz while test-switch is off and taking input from data file. Processor calculation is working after reading the data.
