+++ README.md
@@ -10,6 +10,7 @@ https://github.com/ShubhadeepSadhukhan1993/fastSF
 
 
 ## Installing `fastSF`
+`fastSF` requires several libraries (listed later in this section) for compiling. These libraries can be installed in any directory as long as the said directory is added to the user $PATH. In this section, we will provide the guidelines to install these libraries in $HOME/local (which is generally recommended).
 
 ### Environment Variables
 The following environment variables need to be set before compiling `fastSF`. (You may append these lines to `$HOME/.bashrc`):
@@ -116,44 +117,40 @@ This section provides a detailed procedure to execute `fastSF` for a given veloc
 
 ### i) Files Required and HDF5 Schema:
 
-All the files storing the input fields should be inside the `in` folder. All the input fields and datasets are required to be precisely as per the following schema.
+All the files storing the input fields needs to be in the `hdf5` format and stored inside the `in` folder. All the input datasets are required to be precisely as per the following schema.
 
 #### For two dimensional fields
 
-For vector field, two `hdf5` files are required:
+For vector field, two datasets files are required:
 
-1. The hdf5 file storing the x-component of the vector / velocity field. Unless specified by the user via command line arguments, the hdf5 file should be named as `U.V1r.h5`, with a 2D dataset `U.V1r` of dimensions (`Nx,Nz`) storing real double-precision floating point values. 
-2. The hdf5 file storing the z-component of the vector / velocity field. Unless specified by the user via command line arguments, the hdf5 file should be named as `U.V3r.h5`, with a 2D dataset `U.V3r` of dimensions (`Nx,Nz`) storing real double-precision floating point values. 
+1. A 2D dataset storing the x-component of the vector / velocity field. This dataset has dimensions (`Nx,Nz`) storing real double-precision floating point values. The names of the dataset and the hdf5 file containing the dataset can be arbitrary, but they need to be specified by the user via command line arguments during the execution of `fastSF`. If the user prefers not to use command line arguments, the hdf5 file by default should be named as `U.V1r.h5` and the dataset as `U.V1r`.
 
-For scalar field, For vector field, one `hdf5` file is required:
+2. A 2D dataset storing the z-component of the vector / velocity field. This dataset has dimensions (`Nx,Nz`) storing real double-precision floating point values. The names of the dataset and the hdf5 file containing the dataset can be arbitrary, but they need to be specified by the user via command line arguments during the execution of `fastSF`. If the user prefers not to use command line arguments, the hdf5 file by default should be named as `U.V3r.h5` and the dataset as `U.V3r`. 
 
-1. The hdf5 file storing the scalar field. Unless specified by the user via command line arguments, the hdf5 file should be named as `T.Fr.h5`, with a 2D dataset `T.Fr` of dimensions (`Nx,Nz`) storing real double-precision floating point values.
+For scalar field, one dataset is required:
 
+1. A 2D dataset storing the scalar field. This dataset has dimensions (`Nx,Nz`) storing real double-precision floating point values. The names of the dataset and the hdf5 file containing the dataset can be arbitrary, but they need to be specified by the user via command line arguments during the execution of `fastSF`. If the user prefers not to use command line arguments, the hdf5 file by default should be named as `T.Fr.h5` and the dataset as `T.Fr`.
 
 #### For three dimensional fields
 
 For vector field, three `hdf5` files are required:
 
[…]
