### 2020-12-08 #2185 @shashwat-bhattacharya
Hello @jedbrown
First, thanks a lot for your suggestions. Please accept our apologies for our delay in getting back to you; we are all working remotely under rather suboptimal conditions because of COVID-19 pandemic.

* We agree that asking users to modify Makefile is not a good practice. We wish to clarify that the users will not be required to edit the makefile as long as all the dependencies are installed in any directory that has been added to the user $PATH. We have now clarified this in README.md. Further, we now also mention that the installation instructions in README.md are for installing the dependencies in $HOME/local (which is generally recommended), but the users have the option of installing the dependencies anywhere as long as it is added to the $PATH. Hopefully, this addresses the issue regarding the Makefile.

* We fully agree that fastSF should be compatible with more recent versions hdf5. We have successfully executed it with hdf5 v1.10.5 in Shaheen II of KAUST, Saudi Arabia, thereby showing that it works with newer hdf5 versions. Hence, we have removed the disclaimer regarding hdf5 1.8 from README.md.

* We have modified the code to read the grid dimensions from the hdf5 files. The grid dimensions are now required to be specified only in the testing mode, where the input files are generated by the code and not read from hdf5 files.

* `fastSF` now responds to `-h`, in which the options for command-line arguments are displayed to the user.
