### 2020-12-08 #2185 @shashwat-bhattacharya
Again, thanks for your suggestions @jedbrown .

* We have replaced \mathbf with \boldsymbol everywhere. This should resolve the ambiguity concerning ‘l'.

* “Phys of Fluids” has been corrected to “Physics of Fluids”. Thanks for pointing this out and sorry for this typo!

* We have removed the capitalization in the yaml-cpp reference.

* Thanks a lot for your suggestion in improving the performance of the code. However, currently, it will be difficult to address this as we are all working remotely. However, we will keep upgrading the code and will definitely incorporate your suggestion. We have removed specific claims of high performance its well as the discussion of six vs. three nested loops in the revised manuscript.

Thank you again for your and the referees’ constructive feedback. We feel there has been a significant improvement in the code compared to the version that was first submitted to JOSS thanks to your valuable inputs. We hope that our work is now suitable for publication.
