### 2020-12-21 #2185 @jedbrown
Thanks for your patience. This is great.
* "Structure function is a popular diagnostics tool to study the statistical properties..." -- this grammar is awkward. Perhaps "Structure functions are popular diagnostics in the study of statistical properties ..."
* line 22 : "fastSF is a C++ application for ..." is shorter and reads more clearly.
* I know "ensemble average" is common terminology, but if I understand correctly, what your code computes is a spatial average and that terminology would be more clear (your code isn't averaging across flow snapshots at different times or with different initial conditions).
* I think small figure that plots a structure function (e.g., scalar field in 2D) and the [radial representation](https://github.com/ShubhadeepSadhukhan1993/fastSF/blob/master/docs/Verification.md) (which I think your software doesn't compute, but you might want to ship a script just so users can reproduce standard figures without writing code) would be a lot more informative than the implementation details. Your figure 1 is still useful.
* The yaml-cpp reference looks mangled.
* The h5si reference has a mangled double URL.
* `high-Reynolds` should be `high-{Reynolds}` to protect capitalization
* The hdf5 DOI appears twice

Once you've addressed these points and checked the PDF, it'll be time to archive:
