+++ paper.bib
@@ -204,10 +204,10 @@
 	title = {HDF5},
 	author = {Koziol, Quincey and Robinson, Dana},
 	abstractNote = {HDF5 is a data model, library, and file format for storing and managing data. It supports an unlimited variety of datatypes, and is designed for flexible and efficient I/O and for high volume and complex data. HDF5 is portable and is extensible, allowing applications to evolve in their use of HDF5. The HDF5 Technology suite includes tools and applications for managing, manipulating, viewing, and analyzing data in the HDF5 format.},
-	url = {https://doi.org/10.11578/dc.20180330.1},
 	howpublished = {[Computer Software] \url{https://doi.org/10.11578/dc.20180330.1}},
 	year = {2018},
-	month = {mar}
+	month = {mar},
+	note = {Accessed: 2020-08-03}
 }
 
 @inproceedings{Blitz_web,
@@ -223,10 +223,10 @@
 @misc{H5SI_web,
 	title = {H5SI},
 	author = {Chatterjee, Anando and Chatterjee, Soumyadeep},
-	url = {https://https://github.com/anandogc/h5si},
-	howpublished = {[Computer Software] \url{https://github.com/anandogc/h5si}},
+	howpublished = {[Computer Software] \url {https://https://github.com/anandogc/h5si}},
 	year = {2020},
-	month = {feb}
+	month = {feb},
+	note = {Accessed: 2020-12-21}
 }
 
 @misc{MPICH_Web,
@@ -236,10 +236,11 @@
 }
 
 @article{YAML_web,
-  title={{yaml-cpp,” A YAML parser and emitter for C++,“}},
+  title={{yaml-cpp, A YAML parser and emitter for C++}},
   author={Beder, Jesse},
-  journal={URL: https://github. com/jbeder/yaml-cpp},
-  year={2019}
+  howpublished = {[Computer Software] \url {https://github.com/jbeder/yaml-cpp}},
+  year={2019},
+  note = {Accessed: 2020-08-03}
 }
 
 
