+++ paper.bib
@@ -223,7 +223,7 @@
 @misc{H5SI_web,
 	title = {H5SI},
 	author = {Chatterjee, Anando and Chatterjee, Soumyadeep},
-	howpublished = {[Computer Software] \url {https://https://github.com/anandogc/h5si}},
+	howpublished = {[Computer Software] \url{https://https://github.com/anandogc/h5si}},
 	year = {2020},
 	month = {feb},
 	note = {Accessed: 2020-12-21}
@@ -235,10 +235,10 @@
   note = {Accessed: 2020-08-03}
 }
 
-@article{YAML_web,
+@misc{YAML_web,
   title={{yaml-cpp, A YAML parser and emitter for C++}},
   author={Beder, Jesse},
-  howpublished = {[Computer Software] \url {https://github.com/jbeder/yaml-cpp}},
+  howpublished = {[Computer Software] \url{https://github.com/jbeder/yaml-cpp}},
   year={2019},
   note = {Accessed: 2020-08-03}
 }
