+++ paper.bib
@@ -154,7 +154,7 @@
 	Month = jan,
 	Number = {1},
 	Pages = {165--180},
-	Title = {{Study of high-reynolds number isotropic turbulence by direct numerical simulation}},
+	Title = {{Study of high-Reynolds number isotropic turbulence by direct numerical simulation}},
 	Volume = {41},
 	doi = {10.1146/annurev.fluid.010908.165203},
 	Year = {2009}}
