### 2020-12-23 #2185 @shashwat-bhattacharya
Thanks @jedbrown for your suggestions. I have implemented all of them except for the one about the figure about the structure functions; I have a small confusion here. Should we include the section on [verification](https://github.com/ShubhadeepSadhukhan1993/fastSF/blob/master/docs/Verification.md) in the main paper and shorten the section on the design of the code? Or should we just include a couple of plots for the structure functions of idealized velocity fields with a few lines of explanation in "Velocity and Scalar Structure Functions" section? Thanks a lot for your help.
