### 2020-12-28 #2185 @jedbrown
I think plotting a structure function for a real turbulent flow would be most useful. Verification is important for people who are digging deeper, but I think not the thing to promote up-front.
