### 2021-01-03 0a25a5e
Added SF plot to paper
